Extract AxHost nested classes
