+++ src/System.Windows.Forms/src/System/Windows/Forms/AxHost.cs
@@ -16,7 +16,6 @@ using System.IO;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Runtime.Serialization;
-using System.Runtime.Serialization.Formatters.Binary;
 using System.Threading;
 using System.Windows.Forms.ComponentModel.Com2Interop;
 using System.Windows.Forms.Design;
@@ -33,7 +32,7 @@ namespace System.Windows.Forms
     [DesignTimeVisible(false)]
     [DefaultEvent(nameof(Enter))]
     [Designer("System.Windows.Forms.Design.AxHostDesigner, " + AssemblyRef.SystemDesign)]
-    public abstract class AxHost : Control, ISupportInitialize, ICustomTypeDescriptor
+    public abstract partial class AxHost : Control, ISupportInitialize, ICustomTypeDescriptor
     {
         private static readonly TraceSwitch AxHTraceSwitch = new TraceSwitch("AxHTrace", "ActiveX handle tracing");
         private static readonly TraceSwitch AxPropTraceSwitch = new TraceSwitch("AxPropTrace", "ActiveX property tracing");
@@ -3685,902 +3684,6 @@ namespace System.Windows.Forms
             RaiseCreateHandleEvent(e);
         }
 
-        [AttributeUsage(AttributeTargets.Class, Inherited = false)]
-        public sealed class ClsidAttribute : Attribute
-        {
-            private readonly string val;
-
-            public ClsidAttribute(string clsid)
-            {
-                val = clsid;
-            }
-
-            public string Value
-            {
-                get
-                {
-                    return val;
-                }
-            }
-        }
-
-        [AttributeUsage(AttributeTargets.Assembly, Inherited = false)]
-        public sealed class TypeLibraryTimeStampAttribute : Attribute
-        {
-            private readonly DateTime val;
-
-            public TypeLibraryTimeStampAttribute(string timestamp)
-            {
-                val = DateTime.Parse(timestamp, CultureInfo.InvariantCulture);
-            }
-
-            public DateTime Value
-            {
-                get
-                {
-                    return val;
-                }
-            }
-        }
-
-        public class ConnectionPointCookie
-        {
-            private UnsafeNativeMethods.IConnectionPoint connectionPoint;
-            private int cookie;
-            internal int threadId;
-#if DEBUG
-            private readonly string callStack;
-#endif
-            /// <summary>
-            ///  Creates a connection point to of the given interface type.
-            ///  which will call on a managed code sink that implements that interface.
-            /// </summary>
-            public ConnectionPointCookie(object source, object sink, Type eventInterface)
-                : this(source, sink, eventInterface, true)
-            {
-            }
-
-            internal ConnectionPointCookie(object source, object sink, Type eventInterface, bool throwException)
-            {
-                if (source is UnsafeNativeMethods.IConnectionPointContainer cpc)
-                {
-                    try
-                    {
-                        Guid tmp = eventInterface.GUID;
-                        if (cpc.FindConnectionPoint(ref tmp, out connectionPoint) != NativeMethods.S_OK)
-                        {
-                            connectionPoint = null;
-                        }
-                    }
-                    catch
-                    {
-                        connectionPoint = null;
-                    }
-
-                    if (connectionPoint == null)
-                    {
-                        if (throwException)
-                        {
-                            throw new ArgumentException(string.Format(SR.AXNoEventInterface, eventInterface.Name));
-                        }
-                    }
-                    else if (sink == null || !eventInterface.IsInstanceOfType(sink))
-                    {
-                        if (throwException)
-                        {
-                            throw new InvalidCastException(string.Format(SR.AXNoSinkImplementation, eventInterface.Name));
-                        }
-                    }
-                    else
-                    {
-                        int hr = connectionPoint.Advise(sink, ref cookie);
-                        if (hr == NativeMethods.S_OK)
-                        {
-                            threadId = Thread.CurrentThread.ManagedThreadId;
-                        }
-                        else
-                        {
-                            cookie = 0;
-                            Marshal.ReleaseComObject(connectionPoint);
-                            connectionPoint = null;
-                            if (throwException)
-                            {
-                                throw new InvalidOperationException(string.Format(CultureInfo.CurrentCulture, string.Format(SR.AXNoSinkAdvise, eventInterface.Name), hr));
-                            }
-                        }
-                    }
-                }
-                else
-                {
-                    if (throwException)
-                    {
-                        throw new InvalidCastException(SR.AXNoConnectionPointContainer);
-                    }
-                }
-
-                if (connectionPoint == null || cookie == 0)
-                {
-                    if (connectionPoint != null)
-                    {
-                        Marshal.ReleaseComObject(connectionPoint);
-                    }
-
-                    if (throwException)
-                    {
-                        throw new ArgumentException(string.Format(SR.AXNoConnectionPoint, eventInterface.Name));
-                    }
-                }
-#if DEBUG
-                callStack = Environment.StackTrace;
-#endif
-            }
-
-            /// <summary>
-            ///  Disconnect the current connection point.  If the object is not connected,
-            ///  this method will do nothing.
-            /// </summary>
-            public void Disconnect()
-            {
-                if (connectionPoint != null && cookie != 0)
-                {
-                    try
-                    {
-                        connectionPoint.Unadvise(cookie);
-                    }
-                    catch (Exception ex)
-                    {
-                        if (ClientUtils.IsCriticalException(ex))
-                        {
-                            throw;
-                        }
-                    }
-                    finally
-                    {
-                        cookie = 0;
-                    }
-
-                    try
-                    {
-                        Marshal.ReleaseComObject(connectionPoint);
-                    }
-                    catch (Exception ex)
-                    {
-                        if (ClientUtils.IsCriticalException(ex))
-                        {
-                            throw;
-                        }
-                    }
-                    finally
-                    {
-                        connectionPoint = null;
-                    }
-                }
-            }
-
-            ~ConnectionPointCookie()
-            {
-                if (connectionPoint != null && cookie != 0)
-                {
-                    if (!AppDomain.CurrentDomain.IsFinalizingForUnload())
-                    {
-                        SynchronizationContext context = SynchronizationContext.Current;
-                        context?.Post(new SendOrPostCallback(AttemptDisconnect), null);
-                    }
-                }
-            }
-
-            void AttemptDisconnect(object trash)
-            {
-                if (threadId == Thread.CurrentThread.ManagedThreadId)
-                {
-                    Disconnect();
-                }
-                else
-                {
-                    Debug.Fail("Attempted to disconnect ConnectionPointCookie from the wrong thread (finalizer).");
-                }
-            }
-
-            internal bool Connected
-            {
-                get
-                {
-                    return connectionPoint != null && cookie != 0;
-                }
-            }
-        }
-
-        public enum ActiveXInvokeKind
-        {
-            MethodInvoke,
-            PropertyGet,
-            PropertySet
-        }
-
-        public class InvalidActiveXStateException : Exception
-        {
-            private readonly string name;
-            private readonly ActiveXInvokeKind kind;
-
-            public InvalidActiveXStateException(string name, ActiveXInvokeKind kind)
-            {
-                this.name = name;
-                this.kind = kind;
-            }
-
-            public InvalidActiveXStateException()
-            {
-            }
-
-            public override string ToString()
-            {
-                switch (kind)
-                {
-                    case ActiveXInvokeKind.MethodInvoke:
-                        return string.Format(SR.AXInvalidMethodInvoke, name);
-                    case ActiveXInvokeKind.PropertyGet:
-                        return string.Format(SR.AXInvalidPropertyGet, name);
-                    case ActiveXInvokeKind.PropertySet:
-                        return string.Format(SR.AXInvalidPropertySet, name);
-                    default:
-                        return base.ToString();
-                }
-            }
-        }
-
-        // This private class encapsulates all of the ole interfaces so that users
-        // will not be able to access and call them directly...
-
-        private class OleInterfaces :
-            Ole32.IOleControlSite,
-            UnsafeNativeMethods.IOleClientSite,
-            UnsafeNativeMethods.IOleInPlaceSite,
-            Ole32.ISimpleFrameSite,
-            Ole32.IVBGetControl,
-            Ole32.IGetVBAObject,
-            Ole32.IPropertyNotifySink,
-            IReflect,
-            IDisposable
-        {
-            private readonly AxHost host;
-            private ConnectionPointCookie connectionPoint;
-
-            internal OleInterfaces(AxHost host)
-            {
-                this.host = host ?? throw new ArgumentNullException(nameof(host));
-            }
-
-            private void Dispose(bool disposing)
-            {
-                if (disposing)
-                {
-                    if (!AppDomain.CurrentDomain.IsFinalizingForUnload())
-                    {
-                        SynchronizationContext context = SynchronizationContext.Current;
-                        if (context == null)
-                        {
-                            Debug.Fail("Attempted to disconnect ConnectionPointCookie from the finalizer with no SynchronizationContext.");
-                        }
-                        else
-                        {
-                            context.Post(new SendOrPostCallback(AttemptStopEvents), null);
-                        }
-                    }
-                }
-            }
-
-            public void Dispose()
-            {
-                Dispose(true);
-                GC.SuppressFinalize(this);
-            }
-
-            internal AxHost GetAxHost()
-            {
-                return host;
-            }
-
-            internal void OnOcxCreate()
-            {
-                StartEvents();
-            }
-
-            internal void StartEvents()
-            {
-                if (connectionPoint != null)
-                {
-                    return;
-                }
-
-                object nativeObject = host.GetOcx();
-
-                try
-                {
-                    connectionPoint = new ConnectionPointCookie(nativeObject, this, typeof(Ole32.IPropertyNotifySink));
-                }
-                catch
-                {
-                }
-            }
-
-            void AttemptStopEvents(object trash)
-            {
-                if (connectionPoint == null)
-                {
-                    return;
-                }
-
-                if (connectionPoint.threadId == Thread.CurrentThread.ManagedThreadId)
-                {
-                    StopEvents();
-                }
-                else
-                {
-                    Debug.Fail("Attempted to disconnect ConnectionPointCookie from the wrong thread (finalizer).");
-                }
-            }
-
-            internal void StopEvents()
-            {
-                if (connectionPoint != null)
-                {
-                    connectionPoint.Disconnect();
-                    connectionPoint = null;
-                }
-            }
-
-            // IGetVBAObject methods:
-            unsafe HRESULT Ole32.IGetVBAObject.GetObject(Guid* riid, Ole32.IVBFormat[] rval, uint dwReserved)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in GetObject");
-
-                if (rval == null || riid == null)
-                {
-                    return HRESULT.E_INVALIDARG;
-                }
-
-                if (!riid->Equals(ivbformat_Guid))
-                {
-                    rval[0] = null;
-                    return HRESULT.E_NOINTERFACE;
-                }
-
-                rval[0] = new VBFormat();
-                return HRESULT.S_OK;
-            }
-
-            // IVBGetControl methods:
-
-            unsafe HRESULT Ole32.IVBGetControl.EnumControls(Ole32.OLECONTF dwOleContF, Ole32.GC_WCH dwWhich, out Ole32.IEnumUnknown ppenum)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in EnumControls");
-                ppenum = host.GetParentContainer().EnumControls(host, dwOleContF, dwWhich);
-                return HRESULT.S_OK;
-            }
-
-            // ISimpleFrameSite methods:
-            unsafe HRESULT Ole32.ISimpleFrameSite.PreMessageFilter(IntPtr hwnd, uint msg, IntPtr wp, IntPtr lp, IntPtr* plResult, uint* pdwCookie)
-            {
-                return HRESULT.S_OK;
-            }
-
-            unsafe HRESULT Ole32.ISimpleFrameSite.PostMessageFilter(IntPtr hwnd, uint msg, IntPtr wp, IntPtr lp, IntPtr* plResult, uint dwCookie)
-            {
-                return HRESULT.S_FALSE;
-            }
-
-            // IReflect methods:
-
-            MethodInfo IReflect.GetMethod(string name, BindingFlags bindingAttr, Binder binder, Type[] types, ParameterModifier[] modifiers)
-            {
-                return null;
-            }
-
-            MethodInfo IReflect.GetMethod(string name, BindingFlags bindingAttr)
-            {
-                return null;
-            }
-
-            MethodInfo[] IReflect.GetMethods(BindingFlags bindingAttr)
-            {
-                return Array.Empty<MethodInfo>();
-            }
-
-            FieldInfo IReflect.GetField(string name, BindingFlags bindingAttr)
-            {
-                return null;
-            }
-
-            FieldInfo[] IReflect.GetFields(BindingFlags bindingAttr)
-            {
-                return Array.Empty<FieldInfo>();
-            }
-
-            PropertyInfo IReflect.GetProperty(string name, BindingFlags bindingAttr)
-            {
-                return null;
-            }
-
-            PropertyInfo IReflect.GetProperty(string name, BindingFlags bindingAttr, Binder binder, Type returnType, Type[] types, ParameterModifier[] modifiers)
-            {
-                return null;
-            }
-
-            PropertyInfo[] IReflect.GetProperties(BindingFlags bindingAttr)
-            {
-                return Array.Empty<PropertyInfo>();
-            }
-
-            MemberInfo[] IReflect.GetMember(string name, BindingFlags bindingAttr)
-            {
-                return Array.Empty<MemberInfo>();
-            }
-
-            MemberInfo[] IReflect.GetMembers(BindingFlags bindingAttr)
-            {
-                return Array.Empty<MemberInfo>();
-            }
-
-            object IReflect.InvokeMember(string name, BindingFlags invokeAttr, Binder binder, object target, object[] args, ParameterModifier[] modifiers, CultureInfo culture, string[] namedParameters)
-            {
-                if (name.StartsWith("[DISPID="))
-                {
-                    int endIndex = name.IndexOf(']');
-                    Ole32.DispatchID dispid = (Ole32.DispatchID)int.Parse(name.Substring(8, endIndex - 8), CultureInfo.InvariantCulture);
-                    object ambient = host.GetAmbientProperty(dispid);
-                    if (ambient != null)
-                    {
-                        return ambient;
-                    }
-                }
-
-                throw E_FAIL;
-            }
-
-            Type IReflect.UnderlyingSystemType
-            {
-                get
-                {
-                    return null;
-                }
-            }
-
-            // IOleControlSite methods:
-            HRESULT Ole32.IOleControlSite.OnControlInfoChanged()
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in OnControlInfoChanged");
-                return HRESULT.S_OK;
-            }
-
-            HRESULT Ole32.IOleControlSite.LockInPlaceActive(BOOL fLock)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in LockInPlaceActive");
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT Ole32.IOleControlSite.GetExtendedControl(out object ppDisp)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in GetExtendedControl " + host.ToString());
-                ppDisp = host.GetParentContainer().GetProxyForControl(host);
-                if (ppDisp == null)
-                {
-                    return HRESULT.E_NOTIMPL;
-                }
-
-                return HRESULT.S_OK;
-            }
-
-            unsafe HRESULT Ole32.IOleControlSite.TransformCoords(Point *pPtlHimetric, PointF *pPtfContainer, Ole32.XFORMCOORDS dwFlags)
-            {
-                if (pPtlHimetric == null || pPtfContainer == null)
-                {
-                    return HRESULT.E_INVALIDARG;
-                }
-
-                HRESULT hr = SetupLogPixels(false);
-                if (hr < 0)
-                {
-                    return hr;
-                }
-
-                if ((dwFlags & Ole32.XFORMCOORDS.HIMETRICTOCONTAINER) != 0)
-                {
-                    if ((dwFlags & Ole32.XFORMCOORDS.SIZE) != 0)
-                    {
-                        pPtfContainer->X = (float)host.HM2Pix(pPtlHimetric->X, logPixelsX);
-                        pPtfContainer->Y = (float)host.HM2Pix(pPtlHimetric->Y, logPixelsY);
-                    }
-                    else if ((dwFlags & Ole32.XFORMCOORDS.POSITION) != 0)
-                    {
-                        pPtfContainer->X = (float)host.HM2Pix(pPtlHimetric->X, logPixelsX);
-                        pPtfContainer->Y = (float)host.HM2Pix(pPtlHimetric->Y, logPixelsY);
-                    }
-                    else
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "\t dwFlags not supported: " + dwFlags);
-                        return HRESULT.E_INVALIDARG;
-                    }
-                }
-                else if ((dwFlags & Ole32.XFORMCOORDS.CONTAINERTOHIMETRIC) != 0)
-                {
-                    if ((dwFlags & Ole32.XFORMCOORDS.SIZE) != 0)
-                    {
-                        pPtlHimetric->X = host.Pix2HM((int)pPtfContainer->X, logPixelsX);
-                        pPtlHimetric->Y = host.Pix2HM((int)pPtfContainer->Y, logPixelsY);
-                    }
-                    else if ((dwFlags & Ole32.XFORMCOORDS.POSITION) != 0)
-                    {
-                        pPtlHimetric->X = host.Pix2HM((int)pPtfContainer->X, logPixelsX);
-                        pPtlHimetric->Y = host.Pix2HM((int)pPtfContainer->Y, logPixelsY);
-                    }
-                    else
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "\t dwFlags not supported: " + dwFlags);
-                        return HRESULT.E_INVALIDARG;
-                    }
-                }
-                else
-                {
-                    Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "\t dwFlags not supported: " + dwFlags);
-                    return HRESULT.E_INVALIDARG;
-                }
-
-                return HRESULT.S_OK;
-            }
-
-            unsafe HRESULT Ole32.IOleControlSite.TranslateAccelerator(User32.MSG* pMsg, Ole32.KEYMODIFIERS grfModifiers)
-            {
-                if (pMsg == null)
-                {
-                    return HRESULT.E_POINTER;
-                }
-
-                Debug.Assert(!host.GetAxState(AxHost.siteProcessedInputKey), "Re-entering Ole32.IOleControlSite.TranslateAccelerator!!!");
-                host.SetAxState(AxHost.siteProcessedInputKey, true);
-
-                Message msg = *pMsg;
-                try
-                {
-                    bool f = ((Control)host).PreProcessMessage(ref msg);
-                    return f ? HRESULT.S_OK : HRESULT.S_FALSE;
-                }
-                finally
-                {
-                    host.SetAxState(AxHost.siteProcessedInputKey, false);
-                }
-            }
-
-            HRESULT Ole32.IOleControlSite.OnFocus(BOOL fGotFocus) => HRESULT.S_OK;
-
-            HRESULT Ole32.IOleControlSite.ShowPropertyFrame()
-            {
-                if (host.CanShowPropertyPages())
-                {
-                    host.ShowPropertyPages();
-                    return HRESULT.S_OK;
-                }
-
-                return HRESULT.E_NOTIMPL;
-            }
-
-            // IOleClientSite methods:
-            int UnsafeNativeMethods.IOleClientSite.SaveObject()
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in SaveObject");
-                return NativeMethods.E_NOTIMPL;
-            }
-
-            int UnsafeNativeMethods.IOleClientSite.GetMoniker(int dwAssign, int dwWhichMoniker, out object moniker)
-            {
-                Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetMoniker");
-                moniker = null;
-                return NativeMethods.E_NOTIMPL;
-            }
-
-            HRESULT UnsafeNativeMethods.IOleClientSite.GetContainer(out Ole32.IOleContainer container)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getContainer");
-                container = host.GetParentContainer();
-                return HRESULT.S_OK;
-            }
-
-            unsafe int UnsafeNativeMethods.IOleClientSite.ShowObject()
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in ShowObject");
-                if (host.GetAxState(AxHost.fOwnWindow))
-                {
-                    Debug.Fail("we can't be in showobject if we own our window...");
-                    return NativeMethods.S_OK;
-                }
-                if (host.GetAxState(AxHost.fFakingWindow))
-                {
-                    // we really should not be here...
-                    // this means that the ctl inplace deactivated and didn't call on inplace activate before calling showobject
-                    // so we need to destroy our fake window first...
-                    host.DestroyFakeWindow();
-
-                    // The fact that we have a fake window means that the OCX inplace deactivated when we hid it. It means
-                    // that we have to bring it back from RUNNING to INPLACE so that it can re-create its handle properly.
-                    //
-                    host.TransitionDownTo(OC_LOADED);
-                    host.TransitionUpTo(OC_INPLACE);
-                }
-                if (host.GetOcState() < OC_INPLACE)
-                {
-                    return NativeMethods.S_OK;
-                }
-
-                IntPtr hwnd = IntPtr.Zero;
-                if (host.GetInPlaceObject().GetWindow(&hwnd).Succeeded())
-                {
-                    if (host.GetHandleNoCreate() != hwnd)
-                    {
-                        host.DetachWindow();
-                        if (hwnd != IntPtr.Zero)
-                        {
-                            host.AttachWindow(hwnd);
-                        }
-                    }
-                }
-                else if (host.GetInPlaceObject() is UnsafeNativeMethods.IOleInPlaceObjectWindowless)
-                {
-                    Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "Windowless control.");
-                    throw new InvalidOperationException(SR.AXWindowlessControl);
-                }
-
-                return NativeMethods.S_OK;
-            }
-
-            int UnsafeNativeMethods.IOleClientSite.OnShowWindow(int fShow)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in OnShowWindow");
-                return NativeMethods.S_OK;
-            }
-
-            int UnsafeNativeMethods.IOleClientSite.RequestNewObjectLayout()
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in RequestNewObjectLayout");
-                return NativeMethods.E_NOTIMPL;
-            }
-
-            // IOleInPlaceSite methods:
-
-            unsafe HRESULT UnsafeNativeMethods.IOleInPlaceSite.GetWindow(IntPtr* phwnd)
-            {
-                if (phwnd == null)
-                {
-                    return HRESULT.E_POINTER;
-                }
-
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in GetWindow");
-                Control parent = host.ParentInternal;
-                *phwnd = parent != null ? parent.Handle : IntPtr.Zero;
-                return HRESULT.S_OK;
-            }
-
-            HRESULT UnsafeNativeMethods.IOleInPlaceSite.ContextSensitiveHelp(BOOL fEnterMode)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in ContextSensitiveHelp");
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT UnsafeNativeMethods.IOleInPlaceSite.CanInPlaceActivate()
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in CanInPlaceActivate");
-                return HRESULT.S_OK;
-            }
-
-            int UnsafeNativeMethods.IOleInPlaceSite.OnInPlaceActivate()
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in OnInPlaceActivate");
-                host.SetAxState(AxHost.ownDisposing, false);
-                host.SetAxState(AxHost.rejectSelection, false);
-                host.SetOcState(OC_INPLACE);
-                return NativeMethods.S_OK;
-            }
-
-            int UnsafeNativeMethods.IOleInPlaceSite.OnUIActivate()
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in OnUIActivate for " + host.ToString());
-                host.SetOcState(OC_UIACTIVE);
-                host.GetParentContainer().OnUIActivate(host);
-                return NativeMethods.S_OK;
-            }
-
-            unsafe HRESULT UnsafeNativeMethods.IOleInPlaceSite.GetWindowContext(
-                out Ole32.IOleInPlaceFrame ppFrame,
-                out Ole32.IOleInPlaceUIWindow ppDoc,
-                RECT* lprcPosRect,
-                RECT* lprcClipRect,
-                Ole32.OLEINPLACEFRAMEINFO* lpFrameInfo)
-            {
-                ppDoc = null;
-                ppFrame = host.GetParentContainer();
-
-                if (lprcPosRect == null || lprcClipRect == null)
-                {
-                    return HRESULT.E_POINTER;
-                }
-
-                *lprcPosRect = host.Bounds;
-                *lprcClipRect = WebBrowserHelper.GetClipRect();
-                if (lpFrameInfo != null)
-                {
-                    lpFrameInfo->cb = (uint)Marshal.SizeOf<Ole32.OLEINPLACEFRAMEINFO>();
-                    lpFrameInfo->fMDIApp = BOOL.FALSE;
-                    lpFrameInfo->hAccel = IntPtr.Zero;
-                    lpFrameInfo->cAccelEntries = 0;
-                    lpFrameInfo->hwndFrame = host.ParentInternal.Handle;
-                }
-
-                return HRESULT.S_OK;
-            }
-
-            Interop.HRESULT UnsafeNativeMethods.IOleInPlaceSite.Scroll(Size scrollExtant)
-            {
-                return Interop.HRESULT.S_FALSE;
-            }
-
-            int UnsafeNativeMethods.IOleInPlaceSite.OnUIDeactivate(int fUndoable)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in OnUIDeactivate for " + host.ToString());
-                host.GetParentContainer().OnUIDeactivate(host);
-                if (host.GetOcState() > OC_INPLACE)
-                {
-                    host.SetOcState(OC_INPLACE);
-                }
-                return NativeMethods.S_OK;
-            }
-
-            int UnsafeNativeMethods.IOleInPlaceSite.OnInPlaceDeactivate()
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in OnInPlaceDeactivate");
-                if (host.GetOcState() == OC_UIACTIVE)
-                {
-                    ((UnsafeNativeMethods.IOleInPlaceSite)this).OnUIDeactivate(0);
-                }
-
-                host.GetParentContainer().OnInPlaceDeactivate(host);
-                host.DetachWindow();
-                host.SetOcState(OC_RUNNING);
-                return NativeMethods.S_OK;
-            }
-
-            int UnsafeNativeMethods.IOleInPlaceSite.DiscardUndoState()
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in DiscardUndoState");
-                return NativeMethods.S_OK;
-            }
-
-            HRESULT UnsafeNativeMethods.IOleInPlaceSite.DeactivateAndUndo()
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in DeactivateAndUndo for " + host.ToString());
-                return host.GetInPlaceObject().UIDeactivate();
-            }
-
-            unsafe HRESULT UnsafeNativeMethods.IOleInPlaceSite.OnPosRectChange(RECT* lprcPosRect)
-            {
-                if (lprcPosRect == null)
-                {
-                    return HRESULT.E_INVALIDARG;
-                }
-
-                // The MediaPlayer control has a AllowChangeDisplaySize property that users
-                // can set to control size changes at runtime, but the control itself ignores that and sets the new size.
-                // We prevent this by not allowing controls to call OnPosRectChange(), unless we instantiated the resize.
-                // visual basic6 does the same.
-                //
-                bool useRect = true;
-                if (AxHost.windowsMediaPlayer_Clsid.Equals(host.clsid))
-                {
-                    useRect = host.GetAxState(AxHost.handlePosRectChanged);
-                }
-
-                if (useRect)
-                {
-                    Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in OnPosRectChange" + lprcPosRect->ToString());
-                    RECT clipRect = WebBrowserHelper.GetClipRect();
-                    host.GetInPlaceObject().SetObjectRects(lprcPosRect, &clipRect);
-                    host.MakeDirty();
-                }
-                else
-                {
-                    Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "Control directly called OnPosRectChange... ignoring the new size");
-                }
-
-                return HRESULT.S_OK;
-            }
-
-            // IPropertyNotifySink methods
-
-            HRESULT Ole32.IPropertyNotifySink.OnChanged(Ole32.DispatchID dispid)
-            {
-                // Some controls fire OnChanged() notifications when getting values of some properties.
-                // To prevent this kind of recursion, we check to see if we are already inside a OnChanged() call.
-                if (host.NoComponentChangeEvents != 0)
-                {
-                    return HRESULT.S_OK;
-                }
-
-                host.NoComponentChangeEvents++;
-                try
-                {
-                    AxPropertyDescriptor prop = null;
-
-                    Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in OnChanged");
-
-                    if (dispid != Ole32.DispatchID.UNKNOWN)
-                    {
-                        prop = host.GetPropertyDescriptorFromDispid(dispid);
-                        if (prop != null)
-                        {
-                            prop.OnValueChanged(host);
-                            if (!prop.SettingValue)
-                            {
-                                prop.UpdateTypeConverterAndTypeEditor(true);
-                            }
-                        }
-                    }
-                    else
-                    {
-                        // update them all for DISPID_UNKNOWN.
-                        PropertyDescriptorCollection props = ((ICustomTypeDescriptor)host).GetProperties();
-                        foreach (PropertyDescriptor p in props)
-                        {
-                            prop = p as AxPropertyDescriptor;
-                            if (prop != null && !prop.SettingValue)
-                            {
-                                prop.UpdateTypeConverterAndTypeEditor(true);
-                            }
-                        }
-                    }
-
-                    ISite site = host.Site;
-                    if (site != null)
-                    {
-                        IComponentChangeService changeService = (IComponentChangeService)site.GetService(typeof(IComponentChangeService));
-
-                        if (changeService != null)
-                        {
-                            try
-                            {
-                                changeService.OnComponentChanging(host, prop);
-                            }
-                            catch (CheckoutException coEx)
-                            {
-                                if (coEx == CheckoutException.Canceled)
-                                {
-                                    return HRESULT.S_OK;
-                                }
-                                throw coEx;
-                            }
-
-                            // Now notify the change service that the change was successful.
-                            //
-                            changeService.OnComponentChanged(host, prop, null, (prop?.GetValue(host)));
-                        }
-                    }
-                }
-                catch (Exception t)
-                {
-                    Debug.Fail(t.ToString());
-                    throw t;
-                }
-                finally
-                {
-                    host.NoComponentChangeEvents--;
-                }
-
-                return HRESULT.S_OK;
-            }
-
-            HRESULT Ole32.IPropertyNotifySink.OnRequestEdit(Ole32.DispatchID dispid)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in OnRequestEdit for " + host.ToString());
-                return HRESULT.S_OK;
-            }
-        }
-
         private const int HMperInch = 2540;
         private int Pix2HM(int pix, int logP)
         {
@@ -4875,7 +3978,15 @@ namespace System.Windows.Forms
             return iOleInPlaceActiveObject;
         }
 
-        private UnsafeNativeMethods.IOleObject GetOleObject() => iOleObject ??= (UnsafeNativeMethods.IOleObject)instance;
+        private UnsafeNativeMethods.IOleObject GetOleObject()
+        {
+            if (iOleObject == null)
+            {
+                Debug.Assert(instance != null, "must have the ocx");
+                iOleObject = (UnsafeNativeMethods.IOleObject)instance;
+            }
+            return iOleObject;
+        }
 
         private Ole32.IOleInPlaceObject GetInPlaceObject()
         {
@@ -5349,2582 +4460,6 @@ namespace System.Windows.Forms
             base.OnMouseDown(new MouseEventArgs((MouseButtons)(((int)button) << 20), 1, x, y, 0));
         }
 
-        private class VBFormat : Ole32.IVBFormat
-        {
-            unsafe HRESULT Ole32.IVBFormat.Format(
-                IntPtr vData,
-                IntPtr bstrFormat,
-                IntPtr lpBuffer,
-                ushort cb,
-                int lcid,
-                Ole32.VarFormatFirstDayOfWeek sFirstDayOfWeek,
-                Ole32.VarFormatFirstWeekOfYear sFirstWeekOfYear,
-                ushort* rcb)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in Format");
-                if (rcb == null)
-                {
-                    return HRESULT.E_INVALIDARG;
-                }
-
-                *rcb = 0;
-                if (lpBuffer == IntPtr.Zero || cb < 2)
-                {
-                    return HRESULT.E_INVALIDARG;
-                }
-
-                IntPtr pbstr = IntPtr.Zero;
-                HRESULT hr = Oleaut32.VarFormat(
-                    vData,
-                    bstrFormat,
-                    sFirstDayOfWeek,
-                    sFirstWeekOfYear,
-                    Oleaut32.VarFormatOptions.FORMAT_NOSUBSTITUTE,
-                    ref pbstr);
-                try
-                {
-                    ushort i = 0;
-                    if (pbstr != IntPtr.Zero)
-                    {
-                        short ch = 0;
-                        cb--;
-                        for (; i < cb && (ch = Marshal.ReadInt16(pbstr, i * 2)) != 0; i++)
-                        {
-                            Marshal.WriteInt16(lpBuffer, i * 2, ch);
-                        }
-                    }
-                    Marshal.WriteInt16(lpBuffer, i * 2, (short)0);
-                    *rcb = i;
-                }
-                finally
-                {
-                    Oleaut32.SysFreeString(pbstr);
-                }
-
-                return HRESULT.S_OK;
-            }
-        }
-
-        internal class EnumUnknown : Ole32.IEnumUnknown
-        {
-            private readonly object[] arr;
-            private int loc;
-            private readonly int size;
-
-            internal EnumUnknown(object[] arr)
-            {
-                //if (AxHTraceSwitch.TraceVerbose) Debug.WriteObject(arr);
-                this.arr = arr;
-                loc = 0;
-                size = (arr == null) ? 0 : arr.Length;
-            }
-
-            private EnumUnknown(object[] arr, int loc) : this(arr)
-            {
-                this.loc = loc;
-            }
-
-            unsafe HRESULT Ole32.IEnumUnknown.Next(uint celt, IntPtr rgelt, uint* pceltFetched)
-            {
-                if (pceltFetched != null)
-                {
-                    *pceltFetched = 0;
-                }
-
-                if (celt < 0)
-                {
-                    return HRESULT.E_INVALIDARG;
-                }
-
-                uint fetched = 0;
-                if (loc >= size)
-                {
-                    fetched = 0;
-                }
-                else
-                {
-                    for (; loc < size && fetched < celt; ++loc)
-                    {
-                        if (arr[loc] != null)
-                        {
-                            Marshal.WriteIntPtr(rgelt, Marshal.GetIUnknownForObject(arr[loc]));
-                            rgelt = (IntPtr)((long)rgelt + (long)sizeof(IntPtr));
-                            ++fetched;
-                        }
-                    }
-                }
-
-                if (pceltFetched != null)
-                {
-                    *pceltFetched = fetched;
-                }
-
-                if (fetched != celt)
-                {
-                    return HRESULT.S_FALSE;
-                }
-
-                return HRESULT.S_OK;
-            }
-
-            HRESULT Ole32.IEnumUnknown.Skip(uint celt)
-            {
-                loc += (int)celt;
-                if (loc >= size)
-                {
-                    return HRESULT.S_FALSE;
-                }
-
-                return HRESULT.S_OK;
-            }
-
-            HRESULT Ole32.IEnumUnknown.Reset()
-            {
-                loc = 0;
-                return HRESULT.S_OK;
-            }
-
-            HRESULT Ole32.IEnumUnknown.Clone(out Ole32.IEnumUnknown ppenum)
-            {
-                ppenum = new EnumUnknown(arr, loc);
-                return HRESULT.S_OK;
-            }
-        }
-
-        internal class AxContainer : Ole32.IOleContainer, Ole32.IOleInPlaceFrame, IReflect
-        {
-            internal ContainerControl parent;
-            private IContainer assocContainer; // associated IContainer...
-            // the assocContainer may be null, in which case all this container does is
-            // forward [de]activation messages to the requisite container...
-            private AxHost siteUIActive;
-            private AxHost siteActive;
-            private bool formAlreadyCreated = false;
-            private readonly Hashtable containerCache = new Hashtable();  // name -> Control
-            private int lockCount = 0;
-            private Hashtable components = null;  // Control -> any
-            private Hashtable proxyCache = null;
-            private AxHost ctlInEditMode = null;
-
-            private const int GC_CHILD = 0x1;
-            private const int GC_LASTSIBLING = 0x2;
-            private const int GC_FIRSTSIBLING = 0x4;
-            private const int GC_CONTAINER = 0x20;
-            private const int GC_PREVSIBLING = 0x40;
-            private const int GC_NEXTSIBLING = 0x80;
-
-            internal AxContainer(ContainerControl parent)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in constructor.  Parent created : " + parent.Created.ToString());
-                this.parent = parent;
-                if (parent.Created)
-                {
-                    FormCreated();
-                }
-            }
-
-            // IReflect methods:
-
-            MethodInfo IReflect.GetMethod(string name, BindingFlags bindingAttr, Binder binder, Type[] types, ParameterModifier[] modifiers)
-            {
-                return null;
-            }
-
-            MethodInfo IReflect.GetMethod(string name, BindingFlags bindingAttr)
-            {
-                return null;
-            }
-
-            MethodInfo[] IReflect.GetMethods(BindingFlags bindingAttr)
-            {
-                return Array.Empty<MethodInfo>();
-            }
-
-            FieldInfo IReflect.GetField(string name, BindingFlags bindingAttr)
-            {
-                return null;
-            }
-
-            FieldInfo[] IReflect.GetFields(BindingFlags bindingAttr)
-            {
-                return Array.Empty<FieldInfo>();
-            }
-
-            PropertyInfo IReflect.GetProperty(string name, BindingFlags bindingAttr)
-            {
-                return null;
-            }
-
-            PropertyInfo IReflect.GetProperty(string name, BindingFlags bindingAttr, Binder binder, Type returnType, Type[] types, ParameterModifier[] modifiers)
-            {
-                return null;
-            }
-
-            PropertyInfo[] IReflect.GetProperties(BindingFlags bindingAttr)
-            {
-                return Array.Empty<PropertyInfo>();
-            }
-
-            MemberInfo[] IReflect.GetMember(string name, BindingFlags bindingAttr)
-            {
-                return Array.Empty<MemberInfo>();
-            }
-
-            MemberInfo[] IReflect.GetMembers(BindingFlags bindingAttr)
-            {
-                return Array.Empty<MemberInfo>();
-            }
-
-            object IReflect.InvokeMember(string name, BindingFlags invokeAttr, Binder binder,
-                                                    object target, object[] args, ParameterModifier[] modifiers, CultureInfo culture, string[] namedParameters)
-            {
-
-                foreach (DictionaryEntry e in containerCache)
-                {
-                    string ctlName = GetNameForControl((Control)e.Key);
-                    if (ctlName.Equals(name))
-                    {
-                        return GetProxyForControl((Control)e.Value);
-                    }
-                }
-
-                throw E_FAIL;
-            }
-
-            Type IReflect.UnderlyingSystemType
-            {
-                get
-                {
-                    return null;
-                }
-            }
-
-            internal OleAut32.IExtender GetProxyForControl(Control ctl)
-            {
-                OleAut32.IExtender rval = null;
-                if (proxyCache == null)
-                {
-                    proxyCache = new Hashtable();
-                }
-                else
-                {
-                    rval = (OleAut32.IExtender)proxyCache[ctl];
-                }
-                if (rval == null)
-                {
-                    if (ctl != parent && !GetControlBelongs(ctl))
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "!parent || !belongs NYI");
-                        AxContainer c = FindContainerForControl(ctl);
-                        if (c != null)
-                        {
-                            rval = new ExtenderProxy(ctl, c);
-                        }
-                        else
-                        {
-                            Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "unable to find proxy, returning null");
-                            return null;
-                        }
-                    }
-                    else
-                    {
-                        rval = new ExtenderProxy(ctl, this);
-                    }
-                    proxyCache.Add(ctl, rval);
-                }
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "found proxy " + rval.ToString());
-                return rval;
-            }
-
-            internal string GetNameForControl(Control ctl)
-            {
-                string name = (ctl.Site != null) ? ctl.Site.Name : ctl.Name;
-                return name ?? "";
-            }
-
-            internal void AddControl(Control ctl)
-            {
-                //
-                lock (this)
-                {
-                    if (containerCache.Contains(ctl))
-                    {
-                        throw new ArgumentException(string.Format(SR.AXDuplicateControl, GetNameForControl(ctl)), "ctl");
-                    }
-
-                    containerCache.Add(ctl, ctl);
-
-                    if (assocContainer == null)
-                    {
-                        ISite site = ctl.Site;
-                        if (site != null)
-                        {
-                            assocContainer = site.Container;
-                            IComponentChangeService ccs = (IComponentChangeService)site.GetService(typeof(IComponentChangeService));
-                            if (ccs != null)
-                            {
-                                ccs.ComponentRemoved += new ComponentEventHandler(OnComponentRemoved);
-                            }
-                        }
-                    }
-                    else
-                    {
-#if DEBUG
-                        ISite site = ctl.Site;
-                        if (site != null && assocContainer != site.Container)
-                        {
-                            Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "mismatch between assoc container & added control");
-                        }
-#endif
-                    }
-                }
-            }
-
-            internal void RemoveControl(Control ctl)
-            {
-                //
-                lock (this)
-                {
-                    if (containerCache.Contains(ctl))
-                    {
-                        containerCache.Remove(ctl);
-                    }
-                }
-            }
-
-            private void LockComponents()
-            {
-                lockCount++;
-            }
-
-            private void UnlockComponents()
-            {
-                lockCount--;
-                if (lockCount == 0)
-                {
-                    components = null;
-                }
-            }
-
-            internal Ole32.IEnumUnknown EnumControls(Control ctl, Ole32.OLECONTF dwOleContF, Ole32.GC_WCH dwWhich)
-            {
-                GetComponents();
-
-                LockComponents();
-                try
-                {
-                    ArrayList l = null;
-                    bool selected = (dwWhich & Ole32.GC_WCH.FSELECTED) != 0;
-                    bool reverse = (dwWhich & Ole32.GC_WCH.FREVERSEDIR) != 0;
-                    // Note that visual basic actually ignores the next/prev flags... we will not
-                    bool onlyNext = (dwWhich & Ole32.GC_WCH.FONLYNEXT) != 0;
-                    bool onlyPrev = (dwWhich & Ole32.GC_WCH.FONLYPREV) != 0;
-                    dwWhich &= ~(Ole32.GC_WCH.FSELECTED | Ole32.GC_WCH.FREVERSEDIR |
-                                          Ole32.GC_WCH.FONLYNEXT | Ole32.GC_WCH.FONLYPREV);
-                    if (onlyNext && onlyPrev)
-                    {
-                        Debug.Fail("onlyNext && onlyPrev are both set!");
-                        throw E_INVALIDARG;
-                    }
-                    if (dwWhich == Ole32.GC_WCH.CONTAINER || dwWhich == Ole32.GC_WCH.CONTAINED)
-                    {
-                        if (onlyNext || onlyPrev)
-                        {
-                            Debug.Fail("GC_WCH_FONLYNEXT or FONLYPREV used with CONTANER or CONATINED");
-                            throw E_INVALIDARG;
-                        }
-                    }
-                    int first = 0;
-                    int last = -1; // meaning all
-                    Control[] ctls = null;
-                    switch (dwWhich)
-                    {
-                        default:
-                            Debug.Fail("Bad GC_WCH");
-                            throw E_INVALIDARG;
-                        case Ole32.GC_WCH.CONTAINED:
-                            ctls = ctl.GetChildControlsInTabOrder(false);
-                            ctl = null;
-                            break;
-                        case Ole32.GC_WCH.SIBLING:
-                            Control p = ctl.ParentInternal;
-                            if (p != null)
-                            {
-                                ctls = p.GetChildControlsInTabOrder(false);
-                                if (onlyPrev)
-                                {
-                                    last = ctl.TabIndex;
-                                }
-                                else if (onlyNext)
-                                {
-                                    first = ctl.TabIndex + 1;
-                                }
-                            }
-                            else
-                            {
-                                ctls = Array.Empty<Control>();
-                            }
-                            ctl = null;
-                            break;
-                        case Ole32.GC_WCH.CONTAINER:
-                            l = new ArrayList();
-                            MaybeAdd(l, ctl, selected, dwOleContF, false);
-                            while (ctl != null)
-                            {
-                                AxContainer cont = FindContainerForControl(ctl);
-                                if (cont != null)
-                                {
-                                    MaybeAdd(l, cont.parent, selected, dwOleContF, true);
-                                    ctl = cont.parent;
-                                }
-                                else
-                                {
-                                    break;
-                                }
-                            }
-                            break;
-                        case Ole32.GC_WCH.ALL:
-                            Hashtable htbl = GetComponents();
-                            ctls = new Control[htbl.Keys.Count];
-                            htbl.Keys.CopyTo(ctls, 0);
-                            ctl = parent;
-                            break;
-                    }
-                    if (l == null)
-                    {
-                        l = new ArrayList();
-                        if (last == -1 && ctls != null)
-                        {
-                            last = ctls.Length;
-                        }
-
-                        if (ctl != null)
-                        {
-                            MaybeAdd(l, ctl, selected, dwOleContF, false);
-                        }
-
-                        for (int i = first; i < last; i++)
-                        {
-                            MaybeAdd(l, ctls[i], selected, dwOleContF, false);
-                        }
-                    }
-                    object[] rval = new object[l.Count];
-                    l.CopyTo(rval, 0);
-                    if (reverse)
-                    {
-                        for (int i = 0, j = rval.Length - 1; i < j; i++, j--)
-                        {
-                            object temp = rval[i];
-                            rval[i] = rval[j];
-                            rval[j] = temp;
-                        }
-                    }
-                    return new EnumUnknown(rval);
-                }
-                finally
-                {
-                    UnlockComponents();
-                }
-            }
-
-            private void MaybeAdd(ArrayList l, Control ctl, bool selected, Ole32.OLECONTF dwOleContF, bool ignoreBelong)
-            {
-                if (!ignoreBelong && ctl != parent && !GetControlBelongs(ctl))
-                {
-                    return;
-                }
-
-                if (selected)
-                {
-                    ISelectionService iss = GetSelectionService(ctl);
-                    if (iss == null || !iss.GetComponentSelected(this))
-                    {
-                        return;
-                    }
-                }
-                if (ctl is AxHost hostctl && (dwOleContF & Ole32.OLECONTF.EMBEDDINGS) != 0)
-                {
-                    l.Add(hostctl.GetOcx());
-                }
-                else if ((dwOleContF & Ole32.OLECONTF.OTHERS) != 0)
-                {
-                    object item = GetProxyForControl(ctl);
-                    if (item != null)
-                    {
-                        l.Add(item);
-                    }
-                }
-            }
-
-            private void FillComponentsTable(IContainer container)
-            {
-                if (container != null)
-                {
-                    ComponentCollection comps = container.Components;
-                    if (comps != null)
-                    {
-                        components = new Hashtable();
-                        foreach (IComponent comp in comps)
-                        {
-                            if (comp is Control && comp != parent && comp.Site != null)
-                            {
-                                components.Add(comp, comp);
-                            }
-                        }
-                        return;
-                    }
-                }
-
-                Debug.Assert(parent.Site == null, "Parent is sited but we could not find IContainer!!!");
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "Did not find a container in FillComponentsTable!!!");
-
-                bool checkHashTable = true;
-                Control[] ctls = new Control[containerCache.Values.Count];
-                containerCache.Values.CopyTo(ctls, 0);
-                if (ctls != null)
-                {
-                    if (ctls.Length > 0 && components == null)
-                    {
-                        components = new Hashtable();
-                        checkHashTable = false;
-                    }
-                    for (int i = 0; i < ctls.Length; i++)
-                    {
-                        if (checkHashTable && !components.Contains(ctls[i]))
-                        {
-                            components.Add(ctls[i], ctls[i]);
-                        }
-                    }
-                }
-
-                GetAllChildren(parent);
-            }
-
-            private void GetAllChildren(Control ctl)
-            {
-                if (ctl == null)
-                {
-                    return;
-                }
-
-                if (components == null)
-                {
-                    components = new Hashtable();
-                }
-
-                if (ctl != parent && !components.Contains(ctl))
-                {
-                    components.Add(ctl, ctl);
-                }
-
-                foreach (Control c in ctl.Controls)
-                {
-                    GetAllChildren(c);
-                }
-            }
-
-            private Hashtable GetComponents()
-            {
-                return GetComponents(GetParentsContainer());
-            }
-
-            private Hashtable GetComponents(IContainer cont)
-            {
-                if (lockCount == 0)
-                {
-                    FillComponentsTable(cont);
-                }
-                return components;
-            }
-
-            private bool GetControlBelongs(Control ctl)
-            {
-                Hashtable comps = GetComponents();
-                return comps[ctl] != null;
-            }
-
-            private IContainer GetParentIsDesigned()
-            {
-                ISite site = parent.Site;
-                if (site != null && site.DesignMode)
-                {
-                    return site.Container;
-                }
-
-                return null;
-            }
-
-            private IContainer GetParentsContainer()
-            {
-                IContainer rval = GetParentIsDesigned();
-                Debug.Assert(rval == null || assocContainer == null || (rval == assocContainer),
-                             "mismatch between getIPD & aContainer");
-                return rval ?? assocContainer;
-            }
-
-            private bool RegisterControl(AxHost ctl)
-            {
-                ISite site = ctl.Site;
-                if (site != null)
-                {
-                    IContainer cont = site.Container;
-                    if (cont != null)
-                    {
-                        if (assocContainer != null)
-                        {
-                            return cont == assocContainer;
-                        }
-                        else
-                        {
-                            assocContainer = cont;
-                            IComponentChangeService ccs = (IComponentChangeService)site.GetService(typeof(IComponentChangeService));
-                            if (ccs != null)
-                            {
-                                ccs.ComponentRemoved += new ComponentEventHandler(OnComponentRemoved);
-                            }
-                            return true;
-                        }
-                    }
-                }
-                return false;
-            }
-
-            private void OnComponentRemoved(object sender, ComponentEventArgs e)
-            {
-                if (sender == assocContainer && e.Component is Control c)
-                {
-                    RemoveControl(c);
-                }
-            }
-
-            internal static AxContainer FindContainerForControl(Control ctl)
-            {
-                if (ctl is AxHost axctl)
-                {
-                    if (axctl.container != null)
-                    {
-                        return axctl.container;
-                    }
-
-                    ContainerControl f = axctl.ContainingControl;
-                    if (f != null)
-                    {
-                        AxContainer container = f.CreateAxContainer();
-                        if (container.RegisterControl(axctl))
-                        {
-                            container.AddControl(axctl);
-                            return container;
-                        }
-                    }
-                }
-                return null;
-            }
-
-            internal void OnInPlaceDeactivate(AxHost site)
-            {
-                if (siteActive == site)
-                {
-                    siteActive = null;
-                    if (site.GetSiteOwnsDeactivation())
-                    {
-                        parent.ActiveControl = null;
-                    }
-                    else
-                    {
-                        // we need to tell the form to switch activation to the next thingie...
-                        Debug.Fail("what pathological control is calling inplacedeactivate by itself?");
-                    }
-                }
-            }
-
-            internal void OnUIDeactivate(AxHost site)
-            {
-                Debug.Assert(siteUIActive == null || siteUIActive == site, "deactivating when not active...");
-
-                siteUIActive = null;
-                site.RemoveSelectionHandler();
-                site.SetSelectionStyle(1);
-                site.editMode = EDITM_NONE;
-            }
-
-            internal void OnUIActivate(AxHost site)
-            {
-                // The ShDocVw control repeatedly calls OnUIActivate() with the same
-                // site. This causes the assert below to fire.
-                //
-                if (siteUIActive == site)
-                {
-                    return;
-                }
-
-                if (siteUIActive != null && siteUIActive != site)
-                {
-                    AxHost tempSite = siteUIActive;
-                    bool ownDisposing = tempSite.GetAxState(AxHost.ownDisposing);
-                    try
-                    {
-                        tempSite.SetAxState(AxHost.ownDisposing, true);
-                        tempSite.GetInPlaceObject().UIDeactivate();
-                    }
-                    finally
-                    {
-                        tempSite.SetAxState(AxHost.ownDisposing, ownDisposing);
-                    }
-                }
-                site.AddSelectionHandler();
-                Debug.Assert(siteUIActive == null, "Object did not call OnUIDeactivate");
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "active Object is now " + site.ToString());
-                siteUIActive = site;
-                ContainerControl f = site.ContainingControl;
-                if (f != null)
-                {
-                    f.ActiveControl = site;
-                }
-            }
-
-            private void ListAxControls(ArrayList list, bool fuseOcx)
-            {
-                Hashtable components = GetComponents();
-                if (components == null)
-                {
-                    return;
-                }
-
-                Control[] ctls = new Control[components.Keys.Count];
-                components.Keys.CopyTo(ctls, 0);
-                if (ctls != null)
-                {
-                    for (int i = 0; i < ctls.Length; i++)
-                    {
-                        Control ctl = ctls[i];
-                        if (ctl is AxHost hostctl)
-                        {
-                            if (fuseOcx)
-                            {
-                                list.Add(hostctl.GetOcx());
-                            }
-                            else
-                            {
-                                list.Add(ctl);
-                            }
-                        }
-                    }
-                }
-            }
-
-            internal void ControlCreated(AxHost invoker)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in controlCreated for " + invoker.ToString() + " fAC: " + formAlreadyCreated.ToString());
-                if (formAlreadyCreated)
-                {
-                    if (invoker.IsUserMode() && invoker.AwaitingDefreezing())
-                    {
-                        invoker.Freeze(false);
-                    }
-                }
-                else
-                {
-                    // the form will be created in the future
-                    parent.CreateAxContainer();
-                }
-            }
-
-            internal void FormCreated()
-            {
-                if (formAlreadyCreated)
-                {
-                    return;
-                }
-
-                formAlreadyCreated = true;
-                ArrayList l = new ArrayList();
-                ListAxControls(l, false);
-                AxHost[] axControls = new AxHost[l.Count];
-                l.CopyTo(axControls, 0);
-                for (int i = 0; i < axControls.Length; i++)
-                {
-                    AxHost control = axControls[i];
-                    if (control.GetOcState() >= OC_RUNNING && control.IsUserMode() && control.AwaitingDefreezing())
-                    {
-                        control.Freeze(false);
-                    }
-                }
-            }
-
-            // IOleContainer methods:
-            unsafe HRESULT Ole32.IOleContainer.ParseDisplayName(IntPtr pbc, string pszDisplayName, uint *pchEaten, IntPtr* ppmkOut)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in ParseDisplayName");
-                if (ppmkOut != null)
-                {
-                    *ppmkOut = IntPtr.Zero;
-                }
-
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT Ole32.IOleContainer.EnumObjects(Ole32.OLECONTF grfFlags, out Ole32.IEnumUnknown ppenum)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in EnumObjects");
-                ppenum = null;
-                if ((grfFlags & Ole32.OLECONTF.EMBEDDINGS) != 0)
-                {
-                    Debug.Assert(parent != null, "gotta have it...");
-                    ArrayList list = new ArrayList();
-                    ListAxControls(list, true);
-                    if (list.Count > 0)
-                    {
-                        object[] temp = new object[list.Count];
-                        list.CopyTo(temp, 0);
-                        ppenum = new EnumUnknown(temp);
-                        return HRESULT.S_OK;
-                    }
-                }
-
-                ppenum = new EnumUnknown(null);
-                return HRESULT.S_OK;
-            }
-
-            HRESULT Ole32.IOleContainer.LockContainer(BOOL fLock)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in LockContainer");
-                return HRESULT.E_NOTIMPL;
-            }
-
-            // IOleInPlaceFrame methods:
-            unsafe HRESULT Ole32.IOleInPlaceFrame.GetWindow(IntPtr* phwnd)
-            {
-                if (phwnd == null)
-                {
-                    return HRESULT.E_POINTER;
-                }
-
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in GetWindow");
-                *phwnd = parent.Handle;
-                return HRESULT.S_OK;
-            }
-
-            HRESULT Ole32.IOleInPlaceFrame.ContextSensitiveHelp(BOOL fEnterMode)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in ContextSensitiveHelp");
-                return HRESULT.S_OK;
-            }
-
-            unsafe HRESULT Ole32.IOleInPlaceFrame.GetBorder(RECT* lprectBorder)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in GetBorder");
-                return HRESULT.E_NOTIMPL;
-            }
-
-            unsafe HRESULT Ole32.IOleInPlaceFrame.RequestBorderSpace(RECT* pborderwidths)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in RequestBorderSpace");
-                return HRESULT.E_NOTIMPL;
-            }
-
-            unsafe HRESULT Ole32.IOleInPlaceFrame.SetBorderSpace(RECT* pborderwidths)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in SetBorderSpace");
-                return HRESULT.E_NOTIMPL;
-            }
-
-            internal void OnExitEditMode(AxHost ctl)
-            {
-                Debug.Assert(ctlInEditMode == null || ctlInEditMode == ctl, "who is exiting edit mode?");
-                if (ctlInEditMode == null || ctlInEditMode != ctl)
-                {
-                    return;
-                }
-
-                ctlInEditMode = null;
-            }
-
-            HRESULT Ole32.IOleInPlaceFrame.SetActiveObject(Ole32.IOleInPlaceActiveObject pActiveObject, string pszObjName)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in SetActiveObject " + (pszObjName ?? "<null>"));
-                if (siteUIActive != null)
-                {
-                    if (siteUIActive.iOleInPlaceActiveObjectExternal != pActiveObject)
-                    {
-                        if (siteUIActive.iOleInPlaceActiveObjectExternal != null)
-                        {
-                            Marshal.ReleaseComObject(siteUIActive.iOleInPlaceActiveObjectExternal);
-                        }
-                        siteUIActive.iOleInPlaceActiveObjectExternal = pActiveObject;
-                    }
-                }
-                if (pActiveObject == null)
-                {
-                    if (ctlInEditMode != null)
-                    {
-                        ctlInEditMode.editMode = EDITM_NONE;
-                        ctlInEditMode = null;
-                    }
-                    return HRESULT.S_OK;
-                }
-                AxHost ctl = null;
-                if (pActiveObject is UnsafeNativeMethods.IOleObject oleObject)
-                {
-                    UnsafeNativeMethods.IOleClientSite clientSite = null;
-                    try
-                    {
-                        clientSite = oleObject.GetClientSite();
-                        if (clientSite is OleInterfaces)
-                        {
-                            ctl = ((OleInterfaces)(clientSite)).GetAxHost();
-                        }
-                    }
-                    catch (COMException t)
-                    {
-                        Debug.Fail(t.ToString());
-                    }
-                    if (ctlInEditMode != null)
-                    {
-                        Debug.Fail("control " + ctlInEditMode.ToString() + " did not reset its edit mode to null");
-                        ctlInEditMode.SetSelectionStyle(1);
-                        ctlInEditMode.editMode = EDITM_NONE;
-                    }
-
-                    if (ctl == null)
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "control w/o a valid site called setactiveobject");
-                        ctlInEditMode = null;
-                    }
-                    else
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "resolved to " + ctl.ToString());
-                        if (!ctl.IsUserMode())
-                        {
-                            ctlInEditMode = ctl;
-                            ctl.editMode = EDITM_OBJECT;
-                            ctl.AddSelectionHandler();
-                            ctl.SetSelectionStyle(2);
-                        }
-                    }
-                }
-
-                return HRESULT.S_OK;
-            }
-
-            unsafe HRESULT Ole32.IOleInPlaceFrame.InsertMenus(IntPtr hmenuShared, Ole32.OLEMENUGROUPWIDTHS* lpMenuWidths)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in InsertMenus");
-                return HRESULT.S_OK;
-            }
-
-            HRESULT Ole32.IOleInPlaceFrame.SetMenu(IntPtr hmenuShared, IntPtr holemenu, IntPtr hwndActiveObject)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in SetMenu");
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT Ole32.IOleInPlaceFrame.RemoveMenus(IntPtr hmenuShared)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in RemoveMenus");
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT Ole32.IOleInPlaceFrame.SetStatusText(string pszStatusText)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in SetStatusText");
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT Ole32.IOleInPlaceFrame.EnableModeless(BOOL fEnable)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in EnableModeless");
-                return HRESULT.E_NOTIMPL;
-            }
-
-            unsafe HRESULT Ole32.IOleInPlaceFrame.TranslateAccelerator(User32.MSG* lpmsg, ushort wID)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in IOleInPlaceFrame.TranslateAccelerator");
-                return HRESULT.S_FALSE;
-            }
-
-            // EXPOSED
-
-            private class ExtenderProxy :
-                OleAut32.IExtender,
-                Ole32.IVBGetControl,
-                Ole32.IGetVBAObject,
-                Ole32.IGetOleObject,
-                IReflect
-            {
-                private readonly WeakReference _pRef;
-                private readonly WeakReference _pContainer;
-
-                internal ExtenderProxy(Control principal, AxContainer container)
-                {
-                    _pRef = new WeakReference(principal);
-                    _pContainer = new WeakReference(container);
-                }
-
-                private Control GetP() => (Control)_pRef.Target;
-
-                private AxContainer GetC() => (AxContainer)_pContainer.Target;
-
-                HRESULT Ole32.IVBGetControl.EnumControls(Ole32.OLECONTF dwOleContF, Ole32.GC_WCH dwWhich, out Ole32.IEnumUnknown ppenum)
-                {
-                    Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in EnumControls for proxy");
-                    ppenum = GetC().EnumControls(GetP(), dwOleContF, dwWhich);
-                    return HRESULT.S_OK;
-                }
-
-                unsafe HRESULT Ole32.IGetOleObject.GetOleObject(Guid* riid, out object ppvObj)
-                {
-                    Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in GetOleObject for proxy");
-                    ppvObj = null;
-                    if (riid == null || !riid->Equals(ioleobject_Guid))
-                    {
-                        return HRESULT.E_INVALIDARG;
-                    }
-
-                    if (GetP() is AxHost ctl)
-                    {
-                        ppvObj = ctl.GetOcx();
-                        return HRESULT.S_OK;
-                    }
-
-                    return HRESULT.E_FAIL;
-                }
-
-                unsafe HRESULT Ole32.IGetVBAObject.GetObject(Guid* riid, Ole32.IVBFormat[] rval, uint dwReserved)
-                {
-                    Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in GetObject for proxy");
-                    if (rval == null || riid == null)
-                    {
-                        return HRESULT.E_INVALIDARG;
-                    }
-
-                    if (!riid->Equals(ivbformat_Guid))
-                    {
-                        rval[0] = null;
-                        return HRESULT.E_NOINTERFACE;
-                    }
-
-                    rval[0] = new VBFormat();
-                    return HRESULT.S_OK;
-                }
-
-                public int Align
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getAlign for proxy for " + GetP().ToString());
-                        int rval = (int)((Control)GetP()).Dock;
-                        if (rval < NativeMethods.ActiveX.ALIGN_MIN || rval > NativeMethods.ActiveX.ALIGN_MAX)
-                        {
-                            rval = NativeMethods.ActiveX.ALIGN_NO_CHANGE;
-                        }
-                        return rval;
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setAlign for proxy for " + GetP().ToString() + " " +
-                                          value.ToString(CultureInfo.InvariantCulture));
-                        GetP().Dock = (DockStyle)value;
-                    }
-                }
-
-                public uint BackColor
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getBackColor for proxy for " + GetP().ToString());
-                        return AxHost.GetOleColorFromColor(((Control)GetP()).BackColor);
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setBackColor for proxy for " + GetP().ToString() + " " +
-                                          value.ToString(CultureInfo.InvariantCulture));
-                        GetP().BackColor = AxHost.GetColorFromOleColor(value);
-                    }
-                }
-
-                public BOOL Enabled
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getEnabled for proxy for " + GetP().ToString());
-                        return GetP().Enabled.ToBOOL();
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setEnabled for proxy for " + GetP().ToString() + " " + value.ToString());
-                        GetP().Enabled = value.IsTrue();
-                    }
-                }
-
-                public uint ForeColor
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getForeColor for proxy for " + GetP().ToString());
-                        return AxHost.GetOleColorFromColor(((Control)GetP()).ForeColor);
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setForeColor for proxy for " + GetP().ToString() + " " +
-                                          value.ToString(CultureInfo.InvariantCulture));
-                        GetP().ForeColor = AxHost.GetColorFromOleColor(value);
-                    }
-                }
-
-                public int Height
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getHeight for proxy for " + GetP().ToString());
-                        return Pixel2Twip(GetP().Height, false);
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setHeight for proxy for " + GetP().ToString() + " " +
-                                          Twip2Pixel(value, false).ToString(CultureInfo.InvariantCulture));
-                        GetP().Height = Twip2Pixel(value, false);
-                    }
-                }
-
-                public int Left
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getLeft for proxy for " + GetP().ToString());
-                        return Pixel2Twip(GetP().Left, true);
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setLeft for proxy for " + GetP().ToString() + " " +
-                                          Twip2Pixel(value, true).ToString(CultureInfo.InvariantCulture));
-                        GetP().Left = Twip2Pixel(value, true);
-                    }
-                }
-
-                public object Parent
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getParent for proxy for " + GetP().ToString());
-                        return GetC().GetProxyForControl(GetC().parent);
-                    }
-                }
-
-                public short TabIndex
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getTabIndex for proxy for " + GetP().ToString());
-                        return (short)GetP().TabIndex;
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setTabIndex for proxy for " + GetP().ToString() + " " +
-                                          value.ToString(CultureInfo.InvariantCulture));
-                        GetP().TabIndex = (int)value;
-                    }
-                }
-
-                public BOOL TabStop
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getTabStop for proxy for " + GetP().ToString());
-                        return GetP().TabStop.ToBOOL();
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setTabStop for proxy for " + GetP().ToString() + " " + value.ToString());
-                        GetP().TabStop = value.IsTrue();
-                    }
-                }
-
-                public int Top
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getTop for proxy for " + GetP().ToString());
-                        return Pixel2Twip(GetP().Top, false);
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setTop for proxy for " + GetP().ToString() + " " +
-                                          Twip2Pixel(value, false).ToString(CultureInfo.InvariantCulture));
-                        GetP().Top = Twip2Pixel(value, false);
-                    }
-                }
-
-                public BOOL Visible
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getVisible for proxy for " + GetP().ToString());
-                        return GetP().Visible.ToBOOL();
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setVisible for proxy for " + GetP().ToString() + " " + value.ToString());
-                        GetP().Visible = value.IsTrue();
-                    }
-                }
-
-                public int Width
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getWidth for proxy for " + GetP().ToString());
-                        return Pixel2Twip(GetP().Width, true);
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setWidth for proxy for " + GetP().ToString() + " " +
-                                          Twip2Pixel(value, true).ToString(CultureInfo.InvariantCulture));
-                        GetP().Width = Twip2Pixel(value, true);
-                    }
-                }
-
-                public string Name
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getName for proxy for " + GetP().ToString());
-                        return GetC().GetNameForControl(GetP());
-                    }
-                }
-
-                public IntPtr Hwnd
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getHwnd for proxy for " + GetP().ToString());
-                        return GetP().Handle;
-                    }
-                }
-
-                public object Container => GetC();
-
-                public string Text
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in getText for proxy for " + GetP().ToString());
-                        return GetP().Text;
-                    }
-                    set
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in setText for proxy for " + GetP().ToString());
-                        GetP().Text = value;
-                    }
-                }
-
-                public void Move(object left, object top, object width, object height)
-                {
-                }
-
-                // IReflect methods:
-
-                MethodInfo IReflect.GetMethod(string name, BindingFlags bindingAttr, Binder binder, Type[] types, ParameterModifier[] modifiers)
-                {
-                    return null;
-                }
-
-                MethodInfo IReflect.GetMethod(string name, BindingFlags bindingAttr)
-                {
-                    return null;
-                }
-
-                MethodInfo[] IReflect.GetMethods(BindingFlags bindingAttr)
-                {
-                    return new MethodInfo[] { GetType().GetMethod("Move") };
-                }
-
-                FieldInfo IReflect.GetField(string name, BindingFlags bindingAttr)
-                {
-                    return null;
-                }
-
-                FieldInfo[] IReflect.GetFields(BindingFlags bindingAttr)
-                {
-                    return Array.Empty<FieldInfo>();
-                }
-
-                PropertyInfo IReflect.GetProperty(string name, BindingFlags bindingAttr)
-                {
-                    PropertyInfo prop = GetP().GetType().GetProperty(name, bindingAttr);
-                    if (prop == null)
-                    {
-                        prop = GetType().GetProperty(name, bindingAttr);
-                    }
-                    return prop;
-                }
-
-                PropertyInfo IReflect.GetProperty(string name, BindingFlags bindingAttr, Binder binder, Type returnType, Type[] types, ParameterModifier[] modifiers)
-                {
-                    PropertyInfo prop = GetP().GetType().GetProperty(name, bindingAttr, binder, returnType, types, modifiers);
-                    if (prop == null)
-                    {
-                        prop = GetType().GetProperty(name, bindingAttr, binder, returnType, types, modifiers);
-                    }
-                    return prop;
-                }
-
-                PropertyInfo[] IReflect.GetProperties(BindingFlags bindingAttr)
-                {
-                    PropertyInfo[] extenderProps = GetType().GetProperties(bindingAttr);
-                    PropertyInfo[] ctlProps = GetP().GetType().GetProperties(bindingAttr);
-
-                    if (extenderProps == null)
-                    {
-                        return ctlProps;
-                    }
-                    else if (ctlProps == null)
-                    {
-                        return extenderProps;
-                    }
-                    else
-                    {
-                        int iProp = 0;
-                        PropertyInfo[] props = new PropertyInfo[extenderProps.Length + ctlProps.Length];
-
-                        foreach (PropertyInfo prop in extenderProps)
-                        {
-                            props[iProp++] = prop;
-                        }
-
-                        foreach (PropertyInfo prop in ctlProps)
-                        {
-                            props[iProp++] = prop;
-                        }
-
-                        return props;
-                    }
-                }
-
-                MemberInfo[] IReflect.GetMember(string name, BindingFlags bindingAttr)
-                {
-                    MemberInfo[] memb = GetP().GetType().GetMember(name, bindingAttr);
-                    if (memb == null)
-                    {
-                        memb = GetType().GetMember(name, bindingAttr);
-                    }
-                    return memb;
-                }
-
-                MemberInfo[] IReflect.GetMembers(BindingFlags bindingAttr)
-                {
-                    MemberInfo[] extenderMembs = GetType().GetMembers(bindingAttr);
-                    MemberInfo[] ctlMembs = GetP().GetType().GetMembers(bindingAttr);
-
-                    if (extenderMembs == null)
-                    {
-                        return ctlMembs;
-                    }
-                    else if (ctlMembs == null)
-                    {
-                        return extenderMembs;
-                    }
-                    else
-                    {
-                        MemberInfo[] membs = new MemberInfo[extenderMembs.Length + ctlMembs.Length];
-
-                        Array.Copy(extenderMembs, 0, membs, 0, extenderMembs.Length);
-                        Array.Copy(ctlMembs, 0, membs, extenderMembs.Length, ctlMembs.Length);
-
-                        return membs;
-                    }
-                }
-
-                object IReflect.InvokeMember(string name, BindingFlags invokeAttr, Binder binder,
-                                             object target, object[] args, ParameterModifier[] modifiers, CultureInfo culture, string[] namedParameters)
-                {
-                    try
-                    {
-                        return GetType().InvokeMember(name, invokeAttr, binder, target, args, modifiers, culture, namedParameters);
-                    }
-                    catch (MissingMethodException)
-                    {
-                        return GetP().GetType().InvokeMember(name, invokeAttr, binder, GetP(), args, modifiers, culture, namedParameters);
-                    }
-                }
-
-                Type IReflect.UnderlyingSystemType
-                {
-                    get
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "In UnderlyingSystemType");
-                        return null;
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        ///  StateConverter is a class that can be used to convert
-        ///  State from one data type to another.  Access this
-        ///  class through the TypeDescriptor.
-        /// </summary>
-        public class StateConverter : TypeConverter
-        {
-            /// <summary>
-            ///  Gets a value indicating whether this converter can
-            ///  convert an object in the given source type to the native type of the converter
-            ///  using the context.
-            /// </summary>
-            public override bool CanConvertFrom(ITypeDescriptorContext context, Type sourceType)
-            {
-                if (sourceType == typeof(byte[]))
-                {
-                    return true;
-                }
-
-                return base.CanConvertFrom(context, sourceType);
-            }
-
-            /// <summary>
-            ///  Gets a value indicating whether this converter can
-            ///  convert an object to the given destination type using the context.
-            /// </summary>
-            public override bool CanConvertTo(ITypeDescriptorContext context, Type destinationType)
-            {
-                if (destinationType == typeof(byte[]))
-                {
-                    return true;
-                }
-
-                return base.CanConvertTo(context, destinationType);
-            }
-
-            /// <summary>
-            ///  Converts the given object to the converter's native type.
-            /// </summary>
-            public override object ConvertFrom(ITypeDescriptorContext context, CultureInfo culture, object value)
-            {
-                if (value is byte[])
-                {
-                    MemoryStream ms = new MemoryStream((byte[])value);
-                    return new State(ms);
-                }
-
-                return base.ConvertFrom(context, culture, value);
-            }
-
-            /// <summary>
-            ///  Converts the given object to another type.  The most common types to convert
-            ///  are to and from a string object.  The default implementation will make a call
-            ///  to ToString on the object if the object is valid and if the destination
-            ///  type is string.  If this cannot convert to the desitnation type, this will
-            ///  throw a NotSupportedException.
-            /// </summary>
-            public override object ConvertTo(ITypeDescriptorContext context, CultureInfo culture, object value, Type destinationType)
-            {
-                if (destinationType == null)
-                {
-                    throw new ArgumentNullException(nameof(destinationType));
-                }
-
-                if (destinationType == typeof(byte[]))
-                {
-                    if (value != null)
-                    {
-                        MemoryStream ms = new MemoryStream();
-                        State state = (State)value;
-                        state.Save(ms);
-                        ms.Close();
-                        return ms.ToArray();
-                    }
-                    else
-                    {
-                        return Array.Empty<byte>();
-                    }
-                }
-
-                return base.ConvertTo(context, culture, value, destinationType);
-            }
-        }
-
-        /// <summary>
-        ///  The class which encapsulates the persisted state of the underlying activeX control
-        ///  An instance of this class my be obtained either by calling getOcxState on an
-        ///  AxHost object, or by reading in from a stream.
-        /// </summary>
-        [TypeConverterAttribute(typeof(TypeConverter))]
-        [Serializable] // This exchanges with the native code.
-        public class State : ISerializable
-        {
-            private readonly int VERSION = 1;
-            private int length;
-            private byte[] buffer;
-            internal int type;
-            private MemoryStream ms;
-            private Ole32.IStorage storage;
-            private Ole32.ILockBytes iLockBytes;
-            private bool manualUpdate = false;
-            private string licenseKey = null;
-#pragma warning disable IDE1006
-            private readonly PropertyBagStream PropertyBagBinary; // Do NOT rename (binary serialization).
-#pragma warning restore IDE1006
-
-            // create on save from ipersist stream
-            internal State(MemoryStream ms, int storageType, AxHost ctl, PropertyBagStream propBag)
-            {
-                type = storageType;
-                PropertyBagBinary = propBag;
-                // dangerous?
-                length = (int)ms.Length;
-                this.ms = ms;
-                manualUpdate = ctl.GetAxState(AxHost.manualUpdate);
-                licenseKey = ctl.GetLicenseKey();
-            }
-
-            internal State(PropertyBagStream propBag)
-            {
-                PropertyBagBinary = propBag;
-            }
-
-            internal State(MemoryStream ms)
-            {
-                this.ms = ms;
-                length = (int)ms.Length;
-                InitializeFromStream(ms);
-            }
-
-            // create on init new w/ storage...
-            internal State(AxHost ctl)
-            {
-                CreateStorage();
-                manualUpdate = ctl.GetAxState(AxHost.manualUpdate);
-                licenseKey = ctl.GetLicenseKey();
-                type = STG_STORAGE;
-            }
-
-            public State(Stream ms, int storageType, bool manualUpdate, string licKey)
-            {
-                type = storageType;
-                // dangerous?
-                length = (int)ms.Length;
-                this.manualUpdate = manualUpdate;
-                licenseKey = licKey;
-
-                InitializeBufferFromStream(ms);
-            }
-
-            /**
-             * Constructor used in deserialization
-             */
-            protected State(SerializationInfo info, StreamingContext context)
-            {
-                SerializationInfoEnumerator sie = info.GetEnumerator();
-                if (sie == null)
-                {
-                    return;
-                }
-                for (; sie.MoveNext();)
-                {
-                    if (string.Compare(sie.Name, "Data", true, CultureInfo.InvariantCulture) == 0)
-                    {
-                        try
-                        {
-                            byte[] dat = (byte[])sie.Value;
-                            if (dat != null)
-                            {
-                                InitializeFromStream(new MemoryStream(dat));
-                            }
-
-                        }
-                        catch (Exception e)
-                        {
-                            Debug.Fail("failure: " + e.ToString());
-                        }
-                    }
-                    else if (string.Compare(sie.Name, nameof(PropertyBagBinary), true, CultureInfo.InvariantCulture) == 0)
-                    {
-                        try
-                        {
-                            Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "Loading up property bag from stream...");
-                            byte[] dat = (byte[])sie.Value;
-                            if (dat != null)
-                            {
-                                PropertyBagBinary = new PropertyBagStream();
-                                PropertyBagBinary.Read(new MemoryStream(dat));
-                            }
-
-                        }
-                        catch (Exception e)
-                        {
-                            Debug.Fail("failure: " + e.ToString());
-                        }
-                    }
-                }
-            }
-
-            internal int Type
-            {
-                get
-                {
-                    return type;
-                }
-                set
-                {
-                    type = value;
-                }
-            }
-
-            internal bool _GetManualUpdate()
-            {
-                return manualUpdate;
-            }
-
-            internal string _GetLicenseKey()
-            {
-                return licenseKey;
-            }
-
-            private void CreateStorage()
-            {
-                Debug.Assert(storage == null, "but we already have a storage!!!");
-                IntPtr hglobal = IntPtr.Zero;
-                if (buffer != null)
-                {
-                    hglobal = Kernel32.GlobalAlloc(Kernel32.GMEM.MOVEABLE, (uint)length);
-                    IntPtr pointer = Kernel32.GlobalLock(hglobal);
-                    try
-                    {
-                        if (pointer != IntPtr.Zero)
-                        {
-                            Marshal.Copy(buffer, 0, pointer, length);
-                        }
-                    }
-                    finally
-                    {
-                        Kernel32.GlobalUnlock(hglobal);
-                    }
-                }
-
-                try
-                {
-                    iLockBytes = Ole32.CreateILockBytesOnHGlobal(hglobal, true);
-                    if (buffer == null)
-                    {
-                        storage = Ole32.StgCreateDocfileOnILockBytes(
-                            iLockBytes,
-                            Ole32.STGM.STGM_CREATE | Ole32.STGM.STGM_READWRITE | Ole32.STGM.STGM_SHARE_EXCLUSIVE,
-                            0);
-                    }
-                    else
-                    {
-                        storage = Ole32.StgOpenStorageOnILockBytes(
-                            iLockBytes,
-                            null,
-                            Ole32.STGM.STGM_READWRITE | Ole32.STGM.STGM_SHARE_EXCLUSIVE,
-                            IntPtr.Zero,
-                            0);
-                    }
-                }
-                catch (Exception)
-                {
-                    if (iLockBytes == null && hglobal != IntPtr.Zero)
-                    {
-                        Kernel32.GlobalFree(hglobal);
-                    }
-                    else
-                    {
-                        iLockBytes = null;
-                    }
-
-                    storage = null;
-                }
-            }
-
-            internal UnsafeNativeMethods.IPropertyBag GetPropBag()
-            {
-                return PropertyBagBinary;
-            }
-
-            internal Ole32.IStorage GetStorage()
-            {
-                if (storage == null)
-                {
-                    CreateStorage();
-                }
-
-                return storage;
-            }
-
-            internal Ole32.IStream GetStream()
-            {
-                if (ms == null)
-                {
-                    Debug.Assert(buffer != null, "gotta have the buffer already...");
-                    if (buffer == null)
-                    {
-                        return null;
-                    }
-
-                    ms = new MemoryStream(buffer);
-                }
-                else
-                {
-                    ms.Seek(0, SeekOrigin.Begin);
-                }
-                return new Ole32.GPStream(ms);
-            }
-
-            private void InitializeFromStream(Stream ids)
-            {
-                BinaryReader br = new BinaryReader(ids);
-
-                type = br.ReadInt32();
-                int version = br.ReadInt32();
-                manualUpdate = br.ReadBoolean();
-                int cc = br.ReadInt32();
-                if (cc != 0)
-                {
-                    licenseKey = new string(br.ReadChars(cc));
-                }
-                for (int skipUnits = br.ReadInt32(); skipUnits > 0; skipUnits--)
-                {
-                    int len = br.ReadInt32();
-                    ids.Position += len;
-                }
-
-                length = br.ReadInt32();
-                if (length > 0)
-                {
-                    buffer = br.ReadBytes(length);
-                }
-            }
-
-            private void InitializeBufferFromStream(Stream ids)
-            {
-                BinaryReader br = new BinaryReader(ids);
-
-                length = br.ReadInt32();
-                if (length > 0)
-                {
-                    buffer = br.ReadBytes(length);
-                }
-            }
-
-            internal State RefreshStorage(Ole32.IPersistStorage iPersistStorage)
-            {
-                Debug.Assert(storage != null, "how can we not have a storage object?");
-                Debug.Assert(iLockBytes != null, "how can we have a storage w/o ILockBytes?");
-                if (storage == null || iLockBytes == null)
-                {
-                    return null;
-                }
-
-                iPersistStorage.Save(storage, BOOL.TRUE);
-                storage.Commit(0);
-                iPersistStorage.HandsOffStorage();
-                try
-                {
-                    buffer = null;
-                    ms = null;
-                    iLockBytes.Stat(out Ole32.STATSTG stat, Ole32.STATFLAG.STATFLAG_NONAME);
-                    length = (int)stat.cbSize;
-                    buffer = new byte[length];
-                    IntPtr hglobal = Ole32.GetHGlobalFromILockBytes(iLockBytes);
-                    IntPtr pointer = Kernel32.GlobalLock(hglobal);
-                    try
-                    {
-                        if (pointer != IntPtr.Zero)
-                        {
-                            Marshal.Copy(pointer, buffer, 0, length);
-                        }
-                        else
-                        {
-                            length = 0;
-                            buffer = null;
-                        }
-                    }
-                    finally
-                    {
-                        Kernel32.GlobalUnlock(hglobal);
-                    }
-                }
-                finally
-                {
-                    iPersistStorage.SaveCompleted(storage);
-                }
-                return this;
-            }
-
-            internal void Save(MemoryStream stream)
-            {
-                BinaryWriter bw = new BinaryWriter(stream);
-
-                bw.Write(type);
-                bw.Write(VERSION);
-                bw.Write(manualUpdate);
-                if (licenseKey != null)
-                {
-                    bw.Write(licenseKey.Length);
-                    bw.Write(licenseKey.ToCharArray());
-                }
-                else
-                {
-                    bw.Write((int)0);
-                }
-                bw.Write((int)0); // skip units
-                bw.Write(length);
-                if (buffer != null)
-                {
-                    bw.Write(buffer);
-                }
-                else if (ms != null)
-                {
-                    ms.Position = 0;
-                    ms.WriteTo(stream);
-                }
-                else
-                {
-                    Debug.Assert(length == 0, "if we have no data, then our length has to be 0");
-                }
-            }
-
-            /// <summary>
-            ///  ISerializable private implementation
-            /// </summary>
-            void ISerializable.GetObjectData(SerializationInfo si, StreamingContext context)
-            {
-                MemoryStream stream = new MemoryStream();
-                Save(stream);
-
-                si.AddValue("Data", stream.ToArray());
-
-                if (PropertyBagBinary != null)
-                {
-                    try
-                    {
-                        stream = new MemoryStream();
-                        PropertyBagBinary.Write(stream);
-                        si.AddValue(nameof(PropertyBagBinary), stream.ToArray());
-                    }
-                    catch (Exception e)
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "Failed to serialize the property bag into ResX : " + e.ToString());
-                    }
-                }
-            }
-        }
-
-        internal class PropertyBagStream : UnsafeNativeMethods.IPropertyBag
-        {
-            private Hashtable bag = new Hashtable();
-
-            internal void Read(Stream stream)
-            {
-                BinaryFormatter formatter = new BinaryFormatter();
-                try
-                {
-                    bag = (Hashtable)formatter.Deserialize(stream);
-                }
-                catch
-                {
-                    // Error reading.  Just init an empty hashtable.
-                    bag = new Hashtable();
-                }
-            }
-
-            int UnsafeNativeMethods.IPropertyBag.Read(string pszPropName, ref object pVar, UnsafeNativeMethods.IErrorLog pErrorLog)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "Reading property " + pszPropName + " from OCXState propertybag.");
-
-                if (!bag.Contains(pszPropName))
-                {
-                    return NativeMethods.E_INVALIDARG;
-                }
-
-                pVar = bag[pszPropName];
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "\tValue=" + ((pVar == null) ? "<null>" : pVar.ToString()));
-
-                // The EE returns a VT_EMPTY for a null. The problem is that visual basic6 expects the caller to respect the
-                // "hint" it gives in the VariantType. For eg., for a VT_BSTR, it expects that the callee will null
-                // out the BSTR field of the variant. Since, the EE or us cannot do anything about this, we will return
-                // a E_INVALIDARG rather than let visual basic6 crash.
-                //
-                return (pVar == null) ? NativeMethods.E_INVALIDARG : NativeMethods.S_OK;
-            }
-
-            int UnsafeNativeMethods.IPropertyBag.Write(string pszPropName, ref object pVar)
-            {
-                Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "Writing property " + pszPropName + " [" + pVar + "] into OCXState propertybag.");
-                if (pVar != null && !pVar.GetType().IsSerializable)
-                {
-                    Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "\t " + pVar.GetType().FullName + " is not serializable.");
-                    return NativeMethods.S_OK;
-                }
-
-                bag[pszPropName] = pVar;
-                return NativeMethods.S_OK;
-            }
-
-            internal void Write(Stream stream)
-            {
-                BinaryFormatter formatter = new BinaryFormatter();
-                formatter.Serialize(stream, bag);
-            }
-        }
-
         protected delegate void AboutBoxDelegate();
-
-        [ComVisible(false)]
-        public class AxComponentEditor : WindowsFormsComponentEditor
-        {
-            public override bool EditComponent(ITypeDescriptorContext context, object obj, IWin32Window parent)
-            {
-                if (obj is AxHost host)
-                {
-                    try
-                    {
-                        Debug.WriteLineIf(AxHTraceSwitch.TraceVerbose, "in AxComponentEditor.EditComponent");
-                        ((Ole32.IOleControlSite)host.oleSite).ShowPropertyFrame();
-                        return true;
-                    }
-                    catch (Exception t)
-                    {
-                        Debug.Fail(t.ToString());
-                        throw;
-                    }
-                }
-                return false;
-            }
-        }
-
-        internal class AxPropertyDescriptor : PropertyDescriptor
-        {
-            private readonly PropertyDescriptor baseProp;
-            internal AxHost owner;
-            private readonly DispIdAttribute dispid;
-
-            private TypeConverter converter;
-            private UITypeEditor editor;
-            private readonly ArrayList updateAttrs = new ArrayList();
-            private int flags = 0;
-
-            private const int FlagUpdatedEditorAndConverter = 0x00000001;
-            private const int FlagCheckGetter = 0x00000002;
-            private const int FlagGettterThrew = 0x00000004;
-            private const int FlagIgnoreCanAccessProperties = 0x00000008;
-            private const int FlagSettingValue = 0x00000010;
-
-            internal AxPropertyDescriptor(PropertyDescriptor baseProp, AxHost owner) : base(baseProp)
-            {
-                this.baseProp = baseProp;
-                this.owner = owner;
-
-                // Get the category for this dispid.
-                //
-                dispid = (DispIdAttribute)baseProp.Attributes[typeof(DispIdAttribute)];
-                if (dispid != null)
-                {
-                    // Look to see if this property has a property page.
-                    // If it does, then it needs to be Browsable(true).
-                    //
-                    if (!IsBrowsable && !IsReadOnly)
-                    {
-                        Guid g = GetPropertyPage((Ole32.DispatchID)dispid.Value);
-                        if (!Guid.Empty.Equals(g))
-                        {
-                            Debug.WriteLineIf(AxPropTraceSwitch.TraceVerbose, "Making property: " + Name + " browsable because we found an property page.");
-                            AddAttribute(new BrowsableAttribute(true));
-                        }
-                    }
-
-                    // Use the CategoryAttribute provided by the OCX.
-                    CategoryAttribute cat = owner.GetCategoryForDispid((Ole32.DispatchID)dispid.Value);
-                    if (cat != null)
-                    {
-                        AddAttribute(cat);
-                    }
-
-                    // Check to see if this a DataSource property.
-                    // If it is, we can always get and set the value of this property.
-                    //
-                    if (PropertyType.GUID.Equals(dataSource_Guid))
-                    {
-                        SetFlag(FlagIgnoreCanAccessProperties, true);
-                    }
-                }
-            }
-
-            public override Type ComponentType
-            {
-                get
-                {
-                    return baseProp.ComponentType;
-                }
-            }
-
-            public override TypeConverter Converter
-            {
-                get
-                {
-
-                    if (dispid != null)
-                    {
-                        UpdateTypeConverterAndTypeEditorInternal(false, Dispid);
-                    }
-                    return converter ?? base.Converter;
-                }
-            }
-
-            internal Ole32.DispatchID Dispid
-            {
-                get
-                {
-                    DispIdAttribute dispid = (DispIdAttribute)baseProp.Attributes[typeof(DispIdAttribute)];
-                    if (dispid != null)
-                    {
-                        return (Ole32.DispatchID)dispid.Value;
-                    }
-
-                    return Ole32.DispatchID.UNKNOWN;
-                }
-            }
-
-            public override bool IsReadOnly
-            {
-                get
-                {
-                    return baseProp.IsReadOnly;
-                }
-            }
-
-            public override Type PropertyType
-            {
-                get
-                {
-                    return baseProp.PropertyType;
-                }
-            }
-
-            internal bool SettingValue
-            {
-                get
-                {
-                    return GetFlag(FlagSettingValue);
-                }
-            }
-
-            private void AddAttribute(Attribute attr)
-            {
-                updateAttrs.Add(attr);
-            }
-
-            public override bool CanResetValue(object o)
-            {
-                return baseProp.CanResetValue(o);
-            }
-
-            public override object GetEditor(Type editorBaseType)
-            {
-                UpdateTypeConverterAndTypeEditorInternal(false, (Ole32.DispatchID)dispid.Value);
-
-                if (editorBaseType.Equals(typeof(UITypeEditor)) && editor != null)
-                {
-                    return editor;
-                }
-
-                return base.GetEditor(editorBaseType);
-            }
-
-            private bool GetFlag(int flagValue)
-            {
-                return ((flags & flagValue) == flagValue);
-            }
-
-            private unsafe Guid GetPropertyPage(Ole32.DispatchID dispid)
-            {
-                try
-                {
-                    Ole32.IPerPropertyBrowsing ippb = owner.GetPerPropertyBrowsing();
-                    if (ippb == null)
-                    {
-                        return Guid.Empty;
-                    }
-
-                    Guid rval = Guid.Empty;
-                    if (ippb.MapPropertyToPage(dispid, &rval).Succeeded())
-                    {
-                        return rval;
-                    }
-                }
-                catch (COMException)
-                {
-                }
-                catch (Exception t)
-                {
-                    Debug.Fail(t.ToString());
-                }
-                return Guid.Empty;
-            }
-
-            public override object GetValue(object component)
-            {
-                if ((!GetFlag(FlagIgnoreCanAccessProperties) && !owner.CanAccessProperties) || GetFlag(FlagGettterThrew))
-                {
-                    return null;
-                }
-
-                try
-                {
-                    // Some controls fire OnChanged() notifications when getting values of some properties.
-                    // To prevent this kind of recursion, we check to see if we are already inside a OnChanged() call.
-                    //
-                    owner.NoComponentChangeEvents++;
-                    return baseProp.GetValue(component);
-                }
-                catch (Exception e)
-                {
-                    if (!GetFlag(FlagCheckGetter))
-                    {
-                        Debug.WriteLineIf(AxPropTraceSwitch.TraceVerbose, "Get failed for : " + Name + " with exception: " + e.Message + " .Making property non-browsable.");
-                        SetFlag(FlagCheckGetter, true);
-                        AddAttribute(new BrowsableAttribute(false));
-                        owner.RefreshAllProperties = true;
-                        SetFlag(FlagGettterThrew, true);
-                    }
-                    throw e;
-                }
-                finally
-                {
-                    owner.NoComponentChangeEvents--;
-                }
-            }
-
-            public void OnValueChanged(object component)
-            {
-                OnValueChanged(component, EventArgs.Empty);
-            }
-
-            public override void ResetValue(object o)
-            {
-                baseProp.ResetValue(o);
-            }
-
-            private void SetFlag(int flagValue, bool value)
-            {
-                if (value)
-                {
-                    flags |= flagValue;
-                }
-                else
-                {
-                    flags &= ~flagValue;
-                }
-            }
-
-            public override void SetValue(object component, object value)
-            {
-                if (!GetFlag(FlagIgnoreCanAccessProperties) && !owner.CanAccessProperties)
-                {
-                    return;
-                }
-
-                // State oldOcxState = owner.OcxState;
-
-                try
-                {
-                    SetFlag(FlagSettingValue, true);
-                    if (PropertyType.IsEnum && (value.GetType() != PropertyType))
-                    {
-                        baseProp.SetValue(component, Enum.ToObject(PropertyType, value));
-                    }
-                    else
-                    {
-                        baseProp.SetValue(component, value);
-                    }
-                }
-                finally
-                {
-                    SetFlag(FlagSettingValue, false);
-                }
-
-                OnValueChanged(component);
-                if (owner == component)
-                {
-                    owner.SetAxState(AxHost.valueChanged, true);
-                }
-            }
-
-            public override bool ShouldSerializeValue(object o)
-            {
-                return baseProp.ShouldSerializeValue(o);
-            }
-
-            internal void UpdateAttributes()
-            {
-                if (updateAttrs.Count == 0)
-                {
-                    return;
-                }
-
-                ArrayList attributes = new ArrayList(AttributeArray);
-                foreach (Attribute attr in updateAttrs)
-                {
-                    attributes.Add(attr);
-                }
-
-                Attribute[] temp = new Attribute[attributes.Count];
-                attributes.CopyTo(temp, 0);
-                AttributeArray = temp;
-
-                updateAttrs.Clear();
-            }
-
-            /// <summary>
-            ///  Called externally to update the editor or type converter.
-            ///  This simply sets flags so this will happen, it doesn't actually to the update...
-            ///  we wait and do that on-demand for perf.
-            /// </summary>
-            internal void UpdateTypeConverterAndTypeEditor(bool force)
-            {
-                // if this is an external request, flip the flag to false so we do the update on demand.
-                //
-                if (GetFlag(FlagUpdatedEditorAndConverter) && force)
-                {
-                    SetFlag(FlagUpdatedEditorAndConverter, false);
-                }
-            }
-
-            /// <summary>
-            ///  Called externally to update the editor or type converter.
-            ///  This simply sets flags so this will happen, it doesn't actually to the update...
-            ///  we wait and do that on-demand for perf.
-            /// </summary>
-            internal unsafe void UpdateTypeConverterAndTypeEditorInternal(bool force, Ole32.DispatchID dispid)
-            {
-
-                // check to see if we're being forced here or if the work really
-                // needs to be done.
-                //
-                if (GetFlag(FlagUpdatedEditorAndConverter) && !force)
-                {
-                    return;
-                }
-
-                if (owner.GetOcx() == null)
-                {
-                    return;
-                }
-
-                try
-                {
-                    Ole32.IPerPropertyBrowsing ppb = owner.GetPerPropertyBrowsing();
-
-                    if (ppb != null)
-                    {
-                        bool hasStrings = false;
-
-                        // check for enums
-                        var caStrings = new Ole32.CA_STRUCT();
-                        var caCookies = new Ole32.CA_STRUCT();
-
-                        HRESULT hr = HRESULT.S_OK;
-                        try
-                        {
-                            hr = ppb.GetPredefinedStrings(dispid, &caStrings, &caCookies);
-                        }
-                        catch (ExternalException ex)
-                        {
-                            hr = (HRESULT)ex.ErrorCode;
-                            Debug.Fail("An exception occurred inside IPerPropertyBrowsing::GetPredefinedStrings(dispid=" +
-                                       dispid + "), object type=" + new ComNativeDescriptor().GetClassName(ppb));
-                        }
-
-                        if (hr != HRESULT.S_OK)
-                        {
-                            hasStrings = false;
-                            // Destroy the existing editor if we created the current one
-                            // so if the items have disappeared, we don't hold onto the old
-                            // items.
-                            if (converter is Com2EnumConverter)
-                            {
-                                converter = null;
-                            }
-                        }
-                        else
-                        {
-                            hasStrings = true;
-                        }
-
-                        if (hasStrings)
-                        {
-                            OleStrCAMarshaler stringMarshaler = new OleStrCAMarshaler(caStrings);
-                            Int32CAMarshaler intMarshaler = new Int32CAMarshaler(caCookies);
-
-                            if (stringMarshaler.Count > 0 && intMarshaler.Count > 0)
-                            {
-                                if (converter == null)
-                                {
-                                    converter = new AxEnumConverter(this, new AxPerPropertyBrowsingEnum(this, owner, stringMarshaler, intMarshaler, true));
-                                }
-                                else if (converter is AxEnumConverter)
-                                {
-                                    ((AxEnumConverter)converter).RefreshValues();
-                                    if (((AxEnumConverter)converter).com2Enum is AxPerPropertyBrowsingEnum axEnum)
-                                    {
-                                        axEnum.RefreshArrays(stringMarshaler, intMarshaler);
-                                    }
-
-                                }
-
-                            }
-                            else
-                            {
-                                //hasStrings = false;
-                            }
-                        }
-                        else
-                        {
-                            // if we didn't get any strings, try the proppage edtior
-                            //
-                            // Check to see if this is a property that we have already massaged to be a
-                            // .Net type. If it is, don't bother with custom property pages. We already
-                            // have a .Net Editor for this type.
-                            //
-                            ComAliasNameAttribute comAlias = (ComAliasNameAttribute)baseProp.Attributes[typeof(ComAliasNameAttribute)];
-                            if (comAlias == null)
-                            {
-                                Guid g = GetPropertyPage(dispid);
-
-                                if (!Guid.Empty.Equals(g))
-                                {
-                                    editor = new AxPropertyTypeEditor(this, g);
-
-                                    // Show any non-browsable property that has an editor through a
-                                    // property page.
-                                    //
-                                    if (!IsBrowsable)
-                                    {
-                                        Debug.WriteLineIf(AxPropTraceSwitch.TraceVerbose, "Making property: " + Name + " browsable because we found an editor.");
-                                        AddAttribute(new BrowsableAttribute(true));
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    SetFlag(FlagUpdatedEditorAndConverter, true);
-                }
-                catch (Exception e)
-                {
-                    Debug.WriteLineIf(AxPropTraceSwitch.TraceVerbose, "could not get the type editor for property: " + Name + " Exception: " + e);
-                }
-            }
-        }
-
-        private class AxPropertyTypeEditor : UITypeEditor
-        {
-            private readonly AxPropertyDescriptor propDesc;
-            private Guid guid;
-
-            public AxPropertyTypeEditor(AxPropertyDescriptor pd, Guid guid)
-            {
-                propDesc = pd;
-                this.guid = guid;
-            }
-
-            /// <summary>
-            ///  Takes the value returned from valueAccess.getValue() and modifies or replaces
-            ///  the value, passing the result into valueAccess.setValue().  This is where
-            ///  an editor can launch a modal dialog or create a drop down editor to allow
-            ///  the user to modify the value.  Host assistance in presenting UI to the user
-            ///  can be found through the valueAccess.getService function.
-            /// </summary>
-            public override object EditValue(ITypeDescriptorContext context, IServiceProvider provider, object value)
-            {
-                try
-                {
-                    object instance = context.Instance;
-                    propDesc.owner.ShowPropertyPageForDispid(propDesc.Dispid, guid);
-                }
-                catch (Exception ex1)
-                {
-                    if (provider != null)
-                    {
-                        IUIService uiSvc = (IUIService)provider.GetService(typeof(IUIService));
-                        if (uiSvc != null)
-                        {
-                            uiSvc.ShowError(ex1, SR.ErrorTypeConverterFailed);
-                        }
-                    }
-                }
-                return value;
-            }
-
-            /// <summary>
-            ///  Retrieves the editing style of the Edit method.  If the method
-            ///  is not supported, this will return None.
-            /// </summary>
-            public override UITypeEditorEditStyle GetEditStyle(ITypeDescriptorContext context)
-            {
-                return UITypeEditorEditStyle.Modal;
-            }
-        }
-
-        /// <summary>
-        ///  simple derivation of the com2enumconverter that allows us to intercept
-        ///  the call to GetStandardValues so we can on-demand update the enum values.
-        /// </summary>
-        private class AxEnumConverter : Com2EnumConverter
-        {
-            private readonly AxPropertyDescriptor target;
-
-            public AxEnumConverter(AxPropertyDescriptor target, Com2Enum com2Enum) : base(com2Enum)
-            {
-                this.target = target;
-            }
-
-            public override StandardValuesCollection GetStandardValues(ITypeDescriptorContext context)
-            {
-
-                // make sure the converter has been properly refreshed -- calling
-                // the Converter property does this.
-                //
-                TypeConverter tc = this;
-                tc = target.Converter;
-                return base.GetStandardValues(context);
-
-            }
-
-        }
-
-        // This exists for perf reasons.   We delay doing this until we
-        // are actually asked for the array of values.
-        //
-        private class AxPerPropertyBrowsingEnum : Com2Enum
-        {
-            private readonly AxPropertyDescriptor target;
-            private readonly AxHost owner;
-            private OleStrCAMarshaler nameMarshaller;
-            private Int32CAMarshaler valueMarshaller;
-            private bool arraysFetched;
-
-            public AxPerPropertyBrowsingEnum(AxPropertyDescriptor targetObject, AxHost owner, OleStrCAMarshaler names, Int32CAMarshaler values, bool allowUnknowns) : base(Array.Empty<string>(), Array.Empty<object>(), allowUnknowns)
-            {
-                target = targetObject;
-                nameMarshaller = names;
-                valueMarshaller = values;
-                this.owner = owner;
-                arraysFetched = false;
-            }
-
-            /// <summary>
-            ///  Retrieve a copy of the value array
-            /// </summary>
-            public override object[] Values
-            {
-                get
-                {
-                    EnsureArrays();
-                    return base.Values;
-                }
-            }
-
-            /// <summary>
-            ///  Retrieve a copy of the nme array.
-            /// </summary>
-            public override string[] Names
-            {
-                get
-                {
-                    EnsureArrays();
-                    return base.Names;
-                }
-            }
-
-            // ensure that we have processed the caStructs into arrays
-            // of values and strings
-            //
-            private void EnsureArrays()
-            {
-                if (arraysFetched)
-                {
-                    return;
-                }
-
-                arraysFetched = true;
-
-                try
-                {
-
-                    // marshal the items.
-                    object[] nameItems = nameMarshaller.Items;
-                    object[] cookieItems = valueMarshaller.Items;
-                    Ole32.IPerPropertyBrowsing ppb = (Ole32.IPerPropertyBrowsing)owner.GetPerPropertyBrowsing();
-                    int itemCount = 0;
-
-                    Debug.Assert(cookieItems != null && nameItems != null, "An item array is null");
-
-                    if (nameItems.Length > 0)
-                    {
-                        object[] valueItems = new object[cookieItems.Length];
-                        var var = new Ole32.VARIANT();
-                        int cookie;
-
-                        Debug.Assert(cookieItems.Length == nameItems.Length, "Got uneven names and cookies");
-
-                        // for each name item, we ask the object for it's corresponding value.
-                        //
-                        for (int i = 0; i < nameItems.Length; i++)
-                        {
-                            cookie = (int)cookieItems[i];
-                            if (nameItems[i] == null || !(nameItems[i] is string))
-                            {
-                                Debug.Fail("Bad IPerPropertyBrowsing item [" + i.ToString(CultureInfo.InvariantCulture) + "], name=" + (nameItems == null ? "(unknown)" : nameItems[i].ToString()));
-                                continue;
-                            }
-                            var.vt = (short)Ole32.VARENUM.EMPTY;
-                            HRESULT hr = ppb.GetPredefinedValue(target.Dispid, (uint)cookie, var);
-                            if (hr == HRESULT.S_OK && var.vt != Ole32.VARENUM.EMPTY)
-                            {
-                                valueItems[i] = var.ToObject();
-                            }
-                            var.Clear();
-                            itemCount++;
-                        }
-
-                        // pass this data down to the base Com2Enum object...
-                        if (itemCount > 0)
-                        {
-                            string[] strings = new string[itemCount];
-                            Array.Copy(nameItems, 0, strings, 0, itemCount);
-                            base.PopulateArrays(strings, valueItems);
-                        }
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Debug.Fail("Failed to build IPerPropertyBrowsing editor. " + ex.GetType().Name + ", " + ex.Message);
-                }
-            }
-
-            internal void RefreshArrays(OleStrCAMarshaler names, Int32CAMarshaler values)
-            {
-                nameMarshaller = names;
-                valueMarshaller = values;
-                arraysFetched = false;
-            }
-
-            protected override void PopulateArrays(string[] names, object[] values)
-            {
-                // we call base.PopulateArrays directly when we actually want to do this.
-            }
-
-            public override object FromString(string s)
-            {
-                EnsureArrays();
-                return base.FromString(s);
-            }
-
-            public override string ToString(object v)
-            {
-                EnsureArrays();
-                return base.ToString(v);
-            }
-        }
     }
 }
